==> 3_UKB_brain_structure_prep.Rmd <==
---
title: "UKB_brain_structure_prep"
author: "Hannah Casey"
date: "2022-10-21"
output: html_document
---

```{r setup, include=FALSE}
library(dplyr)
library(utils) #for progress bar in for loop
library(tidyverse)
library(data.table)
library(rlist)
```

## Load in all brains structure IDPs and covariates from UKB.

```{r Load covariate and imaging data}

## Load in formatted UKB dataset containing following covariates and structural imaging data - see 1_duckDB_get_data.R

## Load in data
UKB_imaging_covariates <- read.csv("/Volumes/GenScotDepression/users/hcasey/UKB_brain_structure_CP_MDD/UKB_imaging_covariates.csv", header = T)

## Remove 2nd imaging assessment centre visit from data
UKB_imaging_covariates <- UKB_imaging_covariates %>% select(-contains('.3.0'))

## Recode imaging sites
UKB_imaging_covariates <- UKB_imaging_covariates %>%
  mutate(assessment_centre_first_imaging = recode(assessment_centre_first_imaging,
    `11027` = "Newcastle",
    `11025` = "Cheadle",
    `11026` = "Reading",
    `11028` = "Bristol"
  ))

## Recode sex
UKB_imaging_covariates <- UKB_imaging_covariates %>%
  mutate(sex = recode(sex,
    Male = 0,
    Female = 1
  ))

```

## Calculate Global and lobar meaurments for imaging features

```{r Get global and lobar structural measurments}

## Lobar Cortical Volumes

## Frontal lobe FIDs:
## Superior frontal gyrus = 26815 + 26916
## rostralmiddlefrontal = 26814 + 26915
## Caudalmiddlefrontal = 26791 + 26892
## Pars orbitalis = 26806 + 26907
## pars triangularis = 26807 + 26908
## pars opercularis = 26805 + 26906
## Frontal pole = 26819 + 26920
## Lateral orbitofrontal = 26799 + 26900
## medial orbitofrontal = 26801 + 26902
## Precentral gyrus = 26811 + 26912
## paracentral cortex = 26804 + 26905

## Sum volume of individual structures in frontal lobe
UKB_imaging_covariates$volume_frontal_lobe <- UKB_imaging_covariates$f.26815.2.0 + UKB_imaging_covariates$f.26916.2.0 + UKB_imaging_covariates$f.26814.2.0 +
  UKB_imaging_covariates$f.26915.2.0 + UKB_imaging_covariates$f.26791.2.0 + UKB_imaging_covariates$f.26892.2.0 + UKB_imaging_covariates$f.26806.2.0 +
  UKB_imaging_covariates$f.26907.2.0 + UKB_imaging_covariates$f.26807.2.0 + UKB_imaging_covariates$f.26908.2.0 + UKB_imaging_covariates$f.26805.2.0 + 
  UKB_imaging_covariates$f.26906.2.0 + UKB_imaging_covariates$f.26819.2.0 + UKB_imaging_covariates$f.26920.2.0 + UKB_imaging_covariates$f.26799.2.0 +
  UKB_imaging_covariates$f.26900.2.0 + UKB_imaging_covariates$f.26801.2.0 + UKB_imaging_covariates$f.26902.2.0 + UKB_imaging_covariates$f.26811.2.0 + 
  UKB_imaging_covariates$f.26912.2.0 + UKB_imaging_covariates$f.26804.2.0 + UKB_imaging_covariates$f.26905.2.0

## Temporal lobe FIDs:
## Insula = 26821 + 26922
## Superior temporal = 26817 + 26918
## transverse temporal = 26820 + 26921
## banks STS = 	26789 + 26890
## Middle temporal gyrus = 26802 + 26903
## Inferior temporal gyrus = 26796 + 26897
## Fusiform = 26794 + 26895
## parahippocampal = 26803 + 26904
## entorhinal = 26793 + 26894

## Sum volume of individual structures in temporal lobe
UKB_imaging_covariates$volume_temporal_lobe <- UKB_imaging_covariates$f.26821.2.0 + UKB_imaging_covariates$f.26922.2.0 + UKB_imaging_covariates$f.26817.2.0 +
  UKB_imaging_covariates$f.26918.2.0 + UKB_imaging_covariates$f.26820.2.0 + UKB_imaging_covariates$f.26921.2.0 + UKB_imaging_covariates$f.26789.2.0 +
  UKB_imaging_covariates$f.26890.2.0 + UKB_imaging_covariates$f.26802.2.0 + UKB_imaging_covariates$f.26903.2.0 + UKB_imaging_covariates$f.26796.2.0 +
  UKB_imaging_covariates$f.26897.2.0 + UKB_imaging_covariates$f.26794.2.0 + UKB_imaging_covariates$f.26895.2.0 + UKB_imaging_covariates$f.26803.2.0 +
  UKB_imaging_covariates$f.26904.2.0 + UKB_imaging_covariates$f.26793.2.0 + UKB_imaging_covariates$f.26894.2.0

## Parietal lobe:
## Postcentral gyrus = 26809 + 26910
## paracentral cortex = 26804 + 26905
## Superior parietal cortex = 26816 + 26917
## Inferior parietal cortex = 26795 + 26896
## Supramarginal gyrus = 26818 + 26919
## Precuneus = 26812 + 26913

## Sum volume of individual structures in parietal lobe
UKB_imaging_covariates$volume_parietal_lobe <- UKB_imaging_covariates$f.26809.2.0 + UKB_imaging_covariates$f.26910.2.0 + UKB_imaging_covariates$f.26804.2.0 +
  UKB_imaging_covariates$f.26905.2.0 + UKB_imaging_covariates$f.26816.2.0 + UKB_imaging_covariates$f.26917.2.0 + UKB_imaging_covariates$f.26795.2.0 +
  UKB_imaging_covariates$f.26896.2.0 + UKB_imaging_covariates$f.26818.2.0 + UKB_imaging_covariates$f.26919.2.0 + UKB_imaging_covariates$f.26812.2.0 +
  UKB_imaging_covariates$f.26913.2.0

## Occipital lobe:
## Lateral occipital cortex = 26798, 26899
## Cuneus = 26792, 26893
## Pericalcarine cortex = 26808, 26909
## Lingual gyrus = 26800, 26901

UKB_imaging_covariates$volume_occipital_lobe <- UKB_imaging_covariates$f.26798.2.0 + UKB_imaging_covariates$f.26899.2.0 + UKB_imaging_covariates$f.26792.2.0 +
  UKB_imaging_covariates$f.26893.2.0 + UKB_imaging_covariates$f.26808.2.0 + UKB_imaging_covariates$f.26909.2.0 + UKB_imaging_covariates$f.26800.2.0 +
  UKB_imaging_covariates$f.26901.2.0

## Cingulate lobe:
## Rostral ACC = 26813, 26914
## Caudal ACC = 26790, 26891
## Posterior cingulate cortex = 26810 + 26911
## Cingulate isthmus = 26797 + 26898

UKB_imaging_covariates$volume_cingulate_lobe <- UKB_imaging_covariates$f.26813.2.0 + UKB_imaging_covariates$f.26914.2.0 + UKB_imaging_covariates$f.26790.2.0 +
  UKB_imaging_covariates$f.26891.2.0 + UKB_imaging_covariates$f.26810.2.0 + UKB_imaging_covariates$f.26911.2.0 + UKB_imaging_covariates$f.26797.2.0 +
  UKB_imaging_covariates$f.26898.2.0

## Sum lobar volumes to get global cortical volume
UKB_imaging_covariates$volume_global_cortical <- UKB_imaging_covariates$volume_frontal_lobe + UKB_imaging_covariates$volume_temporal_lobe + UKB_imaging_covariates$volume_parietal_lobe +
  UKB_imaging_covariates$volume_occipital_lobe + UKB_imaging_covariates$volume_cingulate_lobe 



## Calculate cortical lobar area measures
## ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
## Frontal lobe area FIDs:
## Superior frontal gyrus = 26748 + 26849
## rostralmiddlefrontal = 26747 + 26848
## Caudalmiddlefrontal = 26724 + 26825
## Pars orbitalis = 26739 + 26840
## pars triangularis = 26740 + 26841
## pars opercularis = 26805 + 26839
## Frontal pole = 26752 + 26853
## Lateral orbitofrontal = 26732 + 26833
## medial orbitofrontal = 26734 + 26835
## Precentral gyrus = 26744 + 26845
## paracentral cortex = 26737 + 26838

## Sum area of individual structures in frontal lobe
UKB_imaging_covariates$area_frontal_lobe <- UKB_imaging_covariates$f.26748.2.0 + UKB_imaging_covariates$f.26849.2.0 + UKB_imaging_covariates$f.26747.2.0 +
  UKB_imaging_covariates$f.26848.2.0 + UKB_imaging_covariates$f.26724.2.0 + UKB_imaging_covariates$f.26825.2.0 + UKB_imaging_covariates$f.26739.2.0 +
  UKB_imaging_covariates$f.26840.2.0 + UKB_imaging_covariates$f.26740.2.0 + UKB_imaging_covariates$f.26841.2.0 + UKB_imaging_covariates$f.26805.2.0 + 
  UKB_imaging_covariates$f.26839.2.0 + UKB_imaging_covariates$f.26752.2.0 + UKB_imaging_covariates$f.26853.2.0 + UKB_imaging_covariates$f.26732.2.0 +
  UKB_imaging_covariates$f.26833.2.0 + UKB_imaging_covariates$f.26734.2.0 + UKB_imaging_covariates$f.26835.2.0 + UKB_imaging_covariates$f.26744.2.0 + 
  UKB_imaging_covariates$f.26845.2.0 + UKB_imaging_covariates$f.26737.2.0 + UKB_imaging_covariates$f.26838.2.0


## Temporal lobe area FIDs:
## Insula= 26754 + 26855
## Superior temporal = 26750 + 26851
## transverse temporal = 26753 + 26854
## banks STS = 	26722 + 26823
## Middle temporal gyrus = 26735 + 26836
## Inferior temporal gyrus = 26729 + 26730
## Fusiform = 26727 + 26828
## parahippocampal = 26736 + 26837
## entorhinal = 26726 + 26827

## Sum area of individual structures in temporal lobe
UKB_imaging_covariates$area_temporal_lobe <- UKB_imaging_covariates$f.26754.2.0 + UKB_imaging_covariates$f.26855.2.0 + UKB_imaging_covariates$f.26750.2.0 +
  UKB_imaging_covariates$f.26851.2.0 + UKB_imaging_covariates$f.26753.2.0 + UKB_imaging_covariates$f.26854.2.0 + UKB_imaging_covariates$f.26722.2.0 +
  UKB_imaging_covariates$f.26823.2.0 + UKB_imaging_covariates$f.26735.2.0 + UKB_imaging_covariates$f.26836.2.0 + UKB_imaging_covariates$f.26729.2.0 +
  UKB_imaging_covariates$f.26730.2.0 + UKB_imaging_covariates$f.26727.2.0 + UKB_imaging_covariates$f.26828.2.0 + UKB_imaging_covariates$f.26736.2.0 +
  UKB_imaging_covariates$f.26837.2.0 + UKB_imaging_covariates$f.26726.2.0 + UKB_imaging_covariates$f.26827.2.0

## Parietal area lobe:
## Postcentral gyrus = 26742 + 26843
## paracentral cortex = 26737 + 26838
## Superior parietal cortex = 26749 + 26850
## Inferior parietal cortex = 26728 + 26829
## Supramarginal gyrus = 26751 + 26852
## Precuneus = 26745 + 26846

## Sum area of individual structures in parietal lobe
UKB_imaging_covariates$area_parietal_lobe <- UKB_imaging_covariates$f.26742.2.0 + UKB_imaging_covariates$f.26843.2.0 + UKB_imaging_covariates$f.26737.2.0 +
  UKB_imaging_covariates$f.26838.2.0 + UKB_imaging_covariates$f.26749.2.0 + UKB_imaging_covariates$f.26850.2.0 + UKB_imaging_covariates$f.26728.2.0 +
  UKB_imaging_covariates$f.26829.2.0 + UKB_imaging_covariates$f.26751.2.0 + UKB_imaging_covariates$f.26852.2.0 + UKB_imaging_covariates$f.26745.2.0 +
  UKB_imaging_covariates$f.26846.2.0

## Occipital area lobe:
## Lateral occipital cortex = 26731, 26832
## Cuneus = 26725, 26826
## Pericalcarine cortex = 26741, 26842
## Lingual gyrus = 26733, 26834

UKB_imaging_covariates$area_occipital_lobe <- UKB_imaging_covariates$f.26731.2.0 + UKB_imaging_covariates$f.26832.2.0 + UKB_imaging_covariates$f.26725.2.0 +
  UKB_imaging_covariates$f.26826.2.0 + UKB_imaging_covariates$f.26741.2.0 + UKB_imaging_covariates$f.26842.2.0 + UKB_imaging_covariates$f.26733.2.0 +
  UKB_imaging_covariates$f.26834.2.0

## Cingulate area lobe:
## Rostral ACC = 26746, 26847
## Caudal ACC = 26723, 26824
## Posterior cingulate cortex = 26743 + 26844
## Cingulate isthmus = 26730 + 26831

UKB_imaging_covariates$area_cingulate_lobe <- UKB_imaging_covariates$f.26746.2.0 + UKB_imaging_covariates$f.26847.2.0 + UKB_imaging_covariates$f.26723.2.0 +
  UKB_imaging_covariates$f.26824.2.0 + UKB_imaging_covariates$f.26743.2.0 + UKB_imaging_covariates$f.26844.2.0 + UKB_imaging_covariates$f.26730.2.0 +
  UKB_imaging_covariates$f.26831.2.0

## Sum lobar areas to get global cortical area
UKB_imaging_covariates$area_global_cortical <- UKB_imaging_covariates$area_frontal_lobe + UKB_imaging_covariates$area_temporal_lobe + UKB_imaging_covariates$area_parietal_lobe +
  UKB_imaging_covariates$area_occipital_lobe +UKB_imaging_covariates$area_cingulate_lobe

## Calculate cortical lobar thickness measures
## ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
## Frontal lobe thickness FIDs:
## Superior frontal gyrus = 26782 + 26883
## rostralmiddlefrontal = 26781 + 26882
## Caudalmiddlefrontal = 26758 + 26859
## Pars orbitalis = 26773 + 26874
## pars triangularis = 26774 + 26875
## pars opercularis = 26772 + 26873
## Frontal pole = 26786 + 26887
## Lateral orbitofrontal = 26766 + 26867
## medial orbitofrontal = 26768 + 26869
## Precentral gyrus = 26778 + 26879
## paracentral cortex = 26771 + 26872

## Get weighted average of thickness of individual structures in frontal lobe
UKB_imaging_covariates$thickness_frontal_lobe <- ((UKB_imaging_covariates$f.26782.2.0 * UKB_imaging_covariates$f.26748.2.0) + (UKB_imaging_covariates$f.26883.2.0 * UKB_imaging_covariates$f.26849.2.0) + (UKB_imaging_covariates$f.26781.2.0 * UKB_imaging_covariates$f.26747.2.0) + (UKB_imaging_covariates$f.26882.2.0 * UKB_imaging_covariates$f.26848.2.0) + (UKB_imaging_covariates$f.26758.2.0 * UKB_imaging_covariates$f.26724.2.0) + (UKB_imaging_covariates$f.26859.2.0 * UKB_imaging_covariates$f.26825.2.0) + (UKB_imaging_covariates$f.26773.2.0 * UKB_imaging_covariates$f.26739.2.0) + (UKB_imaging_covariates$f.26840.2.0 * UKB_imaging_covariates$f.26874.2.0) + (UKB_imaging_covariates$f.26774.2.0 * UKB_imaging_covariates$f.26740.2.0) + (UKB_imaging_covariates$f.26875.2.0 * UKB_imaging_covariates$f.26841.2.0) + (UKB_imaging_covariates$f.26772.2.0 * UKB_imaging_covariates$f.26805.2.0) + (UKB_imaging_covariates$f.26873.2.0 * UKB_imaging_covariates$f.26839.2.0) + (UKB_imaging_covariates$f.26786.2.0 * UKB_imaging_covariates$f.26752.2.0) + (UKB_imaging_covariates$f.26887.2.0 * UKB_imaging_covariates$f.26853.2.0) + (UKB_imaging_covariates$f.26766.2.0 * UKB_imaging_covariates$f.26732.2.0) + (UKB_imaging_covariates$f.26867.2.0 * UKB_imaging_covariates$f.26833.2.0) + (UKB_imaging_covariates$f.26768.2.0 * UKB_imaging_covariates$f.26734.2.0) + (UKB_imaging_covariates$f.26869.2.0 * UKB_imaging_covariates$f.26835.2.0)+ (UKB_imaging_covariates$f.26778.2.0 * UKB_imaging_covariates$f.26744.2.0) + (UKB_imaging_covariates$f.26879.2.0 * UKB_imaging_covariates$f.26845.2.0) + (UKB_imaging_covariates$f.26771.2.0 * UKB_imaging_covariates$f.26737.2.0) + (UKB_imaging_covariates$f.26872.2.0 * UKB_imaging_covariates$f.26838.2.0)) / UKB_imaging_covariates$area_frontal_lobe

## Temporal lobe thickness FIDs:
## Insula= 26788 + 26889
## Superior temporal = 26784 + 26885
## transverse temporal = 26787 + 26888
## banks STS = 	26756 + 26857
## Middle temporal gyrus = 26769 + 26870
## Inferior temporal gyrus = 26763 + 26764
## Fusiform = 26761 + 26862
## parahippocampal = 26770 + 26871
## entorhinal = 26760 + 26760

## Sum thickness of individual structures in temporal lobe
UKB_imaging_covariates$thickness_temporal_lobe <- ((UKB_imaging_covariates$f.26788.2.0 * UKB_imaging_covariates$f.26754.2.0 ) + (UKB_imaging_covariates$f.26889.2.0 * UKB_imaging_covariates$f.26855.2.0 ) + (UKB_imaging_covariates$f.26784.2.0 * UKB_imaging_covariates$f.26750.2.0 ) + (UKB_imaging_covariates$f.26885.2.0 * UKB_imaging_covariates$f.26851.2.0 ) + (UKB_imaging_covariates$f.26787.2.0 * UKB_imaging_covariates$f.26753.2.0) + (UKB_imaging_covariates$f.26888.2.0 * UKB_imaging_covariates$f.26854.2.0) + (UKB_imaging_covariates$f.26756.2.0 * UKB_imaging_covariates$f.26722.2.0) + (UKB_imaging_covariates$f.26857.2.0 * UKB_imaging_covariates$f.26823.2.0) + (UKB_imaging_covariates$f.26769.2.0 * UKB_imaging_covariates$f.26735.2.0) + (UKB_imaging_covariates$f.26870.2.0 * UKB_imaging_covariates$f.26836.2.0) + (UKB_imaging_covariates$f.26763.2.0 * UKB_imaging_covariates$f.26729.2.0) + (UKB_imaging_covariates$f.26764.2.0 * UKB_imaging_covariates$f.26730.2.0) + (UKB_imaging_covariates$f.26761.2.0 * UKB_imaging_covariates$f.26727.2.0) + (UKB_imaging_covariates$f.26862.2.0 * UKB_imaging_covariates$f.26828.2.0) + (UKB_imaging_covariates$f.26770.2.0 * UKB_imaging_covariates$f.26736.2.0) + (UKB_imaging_covariates$f.26871.2.0 * UKB_imaging_covariates$f.26837.2.0) + (UKB_imaging_covariates$f.26760.2.0 * UKB_imaging_covariates$f.26726.2.0) + (UKB_imaging_covariates$f.26760.2.0 * UKB_imaging_covariates$f.26827.2.0)) / UKB_imaging_covariates$area_temporal_lobe

## Parietal thickness lobe:
## Postcentral gyrus = 26776 + 26877
## paracentral cortex = 26771 + 26872
## Superior parietal cortex = 26783 + 26884
## Inferior parietal cortex = 26762 + 26863
## Supramarginal gyrus = 26785 + 26886
## Precuneus = 26779 + 26880

## Sum thickness of individual structures in parietal lobe
UKB_imaging_covariates$thickness_parietal_lobe <- ((UKB_imaging_covariates$f.26776.2.0 * UKB_imaging_covariates$f.26742.2.0) + (UKB_imaging_covariates$f.26877.2.0 * UKB_imaging_covariates$f.26843.2.0) + (UKB_imaging_covariates$f.26771.2.0 * UKB_imaging_covariates$f.26737.2.0) + (UKB_imaging_covariates$f.26872.2.0 * UKB_imaging_covariates$f.26838.2.0) + (UKB_imaging_covariates$f.26783.2.0 * UKB_imaging_covariates$f.26749.2.0) + (UKB_imaging_covariates$f.26884.2.0 * UKB_imaging_covariates$f.26850.2.0) + (UKB_imaging_covariates$f.26762.2.0 * UKB_imaging_covariates$f.26728.2.0) + (UKB_imaging_covariates$f.26863.2.0 * UKB_imaging_covariates$f.26829.2.0) + (UKB_imaging_covariates$f.26785.2.0 * UKB_imaging_covariates$f.26751.2.0) + (UKB_imaging_covariates$f.26886.2.0 * UKB_imaging_covariates$f.26852.2.0) + (UKB_imaging_covariates$f.26779.2.0 * UKB_imaging_covariates$f.26745.2.0) + (UKB_imaging_covariates$f.26880.2.0 * UKB_imaging_covariates$f.26846.2.0)) / UKB_imaging_covariates$area_parietal_lobe

## Occipital thickness lobe:
## Lateral occipital cortex = 26765, 26866
## Cuneus = 26759, 26860
## Pericalcarine cortex = 26775, 26876
## Lingual gyrus = 26767, 26868

UKB_imaging_covariates$thickness_occipital_lobe <- ((UKB_imaging_covariates$f.26765.2.0 * UKB_imaging_covariates$f.26731.2.0) + (UKB_imaging_covariates$f.26866.2.0 * UKB_imaging_covariates$f.26832.2.0) + (UKB_imaging_covariates$f.26759.2.0 * UKB_imaging_covariates$f.26725.2.0) + (UKB_imaging_covariates$f.26860.2.0 * UKB_imaging_covariates$f.26826.2.0) + (UKB_imaging_covariates$f.26775.2.0 * UKB_imaging_covariates$f.26741.2.0) + (UKB_imaging_covariates$f.26876.2.0 * UKB_imaging_covariates$f.26842.2.0) + (UKB_imaging_covariates$f.26767.2.0 * UKB_imaging_covariates$f.26733.2.0) + (UKB_imaging_covariates$f.26868.2.0 * UKB_imaging_covariates$f.26834.2.0)) / UKB_imaging_covariates$area_occipital_lobe

## Cingulate thickness lobe:
## Rostral ACC = 26780, 26881
## Caudal ACC = 26757, 26858
## Posterior cingulate cortex = 26777 + 26878
## Cingulate isthmus = 26764 + 26865

UKB_imaging_covariates$thickness_cingulate_lobe <- ((UKB_imaging_covariates$f.26780.2.0 * UKB_imaging_covariates$f.26746.2.0) + (UKB_imaging_covariates$f.26881.2.0 * UKB_imaging_covariates$f.26847.2.0) + (UKB_imaging_covariates$f.26757.2.0 * UKB_imaging_covariates$f.26723.2.0) + (UKB_imaging_covariates$f.26858.2.0 * UKB_imaging_covariates$f.26824.2.0) + (UKB_imaging_covariates$f.26777.2.0 * UKB_imaging_covariates$f.26743.2.0) + (UKB_imaging_covariates$f.26878.2.0 * UKB_imaging_covariates$f.26844.2.0) + (UKB_imaging_covariates$f.26764.2.0 * UKB_imaging_covariates$f.26730.2.0) + (UKB_imaging_covariates$f.26865.2.0 * UKB_imaging_covariates$f.26831.2.0)) / UKB_imaging_covariates$area_cingulate_lobe

## Sum lobar thicknesss to get global cortical thickness
UKB_imaging_covariates$thickness_global_cortical <- ((UKB_imaging_covariates$thickness_frontal_lobe * UKB_imaging_covariates$area_frontal_lobe) + (UKB_imaging_covariates$thickness_temporal_lobe * UKB_imaging_covariates$area_temporal_lobe) + (UKB_imaging_covariates$thickness_parietal_lobe * UKB_imaging_covariates$area_parietal_lobe) + (UKB_imaging_covariates$thickness_occipital_lobe * UKB_imaging_covariates$area_occipital_lobe) + (UKB_imaging_covariates$thickness_cingulate_lobe * UKB_imaging_covariates$area_cingulate_lobe)) / UKB_imaging_covariates$area_global_cortical


## Calculate general MD measures
## ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
## Association/Commissural fibres

# Cingulate gyrus part of cingulum (left) - f.25519
# Cingulate gyrus part of cingulum (right) - f.25520
# Inferior fronto-occipital fasciculus (left) - f.25527
# Inferior fronto-occipital fasciculus (right) - f.25528
# Inferior longitudinal fasciculus (left) - f.25529
# Inferior longitudinal fasciculus (right) - f.25530
# Parahippocampal part of cingulum (left) - f.25521
# Parahippocampal part of cingulum (right) - f.25522
# Superior longitudinal fasciculus (left) - f.25536
# Superior longitudinal fasciculus (right) - f.25537
# Uncinate fasciculus (left) - f.25540
# Uncinate fasciculus (right) - f.25541
# Forceps major - f.25525
# Forceps minor - f.25526

MD_association_fibres <- c("f.25519.2.0", "f.25520.2.0", "f.25527.2.0", "f.25528.2.0", "f.25529.2.0", "f.25530.2.0",
                           "f.25521.2.0", "f.25522.2.0", "f.25536.2.0", "f.25537.2.0", "f.25540.2.0", "f.25541.2.0",
                           "f.25525.2.0", "f.25526.2.0")


## Thalamic radiations

# Anterior thalamic radiation (left)  - f.25517
# Anterior thalamic radiation (right) - f.25518
# Posterior thalamic radiation (left) - f.25534
# Posterior thalamic radiation (right) - f.25535
# Superior thalamic radiation (left) - f.25538
# Superior thalamic radiation (right) - f.25539

MD_thalamic_radiations <- c("f.25517.2.0", "f.25518.2.0", "f.25534.2.0", "f.25535.2.0", "f.25538.2.0", "f.25539.2.0")


## Projection fibres

# Acoustic radiation (left) - f.25516
# Acoustic radiation (right) - f.25517
# Corticospinal tract (left) - f.25523
# Corticospinal tract (right) - f.25524
# Medial lemniscus (left) - f.25532
# Medial lemniscus (right) - f.25533
# Middle cerebellar peduncle - f.25531

MD_projection_fibres <- c("f.25516.2.0", "f.25517.2.0", "f.25523.2.0", "f.25524.2.0", "f.25532.2.0", "f.25533.2.0", "f.25531.2.0")

## MD_global
MD_global <- c(MD_association_fibres, MD_thalamic_radiations, MD_projection_fibres)

## Remove participants without DTI information and store in temporary file (PCA does not allow NAs)
UKB_imaging_covariates_temp <- UKB_imaging_covariates[!is.na(UKB_imaging_covariates$f.25516.2.0),]

## Run PCA in each general MD measure
PC_res_MD_association_fibres <- princomp((UKB_imaging_covariates_temp[,MD_association_fibres]),scores = TRUE)$scores[,1]

PC_res_MD_association_fibres_PCA <- princomp(UKB_imaging_covariates_temp[,MD_association_fibres])
temp <- PC_res_MD_association_fibres_PCA$sdev^2 / sum(PC_res_MD_association_fibres_PCA$sdev^2)
PC_res_MD_association_fibres_VE <- temp[1]
                                             
PC_res_MD_thalamic_radiations <- princomp(UKB_imaging_covariates_temp[,MD_thalamic_radiations],scores = TRUE)$scores[,1]

PC_res_MD_thalamic_radiations_PCA <- princomp(UKB_imaging_covariates_temp[,MD_thalamic_radiations])
temp <- PC_res_MD_thalamic_radiations_PCA$sdev^2 / sum(PC_res_MD_thalamic_radiations_PCA$sdev^2)
PC_res_MD_thalamic_radiations_VE <- temp[1]

PC_res_MD_projection_fibres <- princomp(UKB_imaging_covariates_temp[,MD_projection_fibres],scores = TRUE)$scores[,1]

PC_res_MD_projection_fibres_PCA <- princomp(UKB_imaging_covariates_temp[,MD_projection_fibres])
temp <- PC_res_MD_projection_fibres_PCA$sdev^2 / sum(PC_res_MD_projection_fibres_PCA$sdev^2)
PC_res_MD_projection_fibres_VE <- temp[1]

PC_res_MD_global <- princomp(UKB_imaging_covariates_temp[,MD_global],scores = TRUE)$scores[,1]

PC_res_MD_global_PCA <- princomp(UKB_imaging_covariates_temp[,MD_global])
temp <- PC_res_MD_global_PCA$sdev^2 / sum(PC_res_MD_global_PCA$sdev^2)
PC_res_MD_global_VE <- temp[1]

## Add 1st PC to temporary dataframe
UKB_imaging_covariates_temp <- cbind(as.data.frame(UKB_imaging_covariates_temp[,"f.eid"]), PC_res_MD_association_fibres, PC_res_MD_thalamic_radiations, PC_res_MD_projection_fibres, PC_res_MD_global)

## Rename column containing 1st PC of each general MD meaure and keep only ID and general MD meaures
UKB_imaging_covariates_temp <- UKB_imaging_covariates_temp %>%
  rename(f.eid = `UKB_imaging_covariates_temp[, "f.eid"]`,
         MD_global= `PC_res_MD_global`,
         MD_association_fibres= `PC_res_MD_association_fibres`,
         MD_thalamic_radiations= `PC_res_MD_thalamic_radiations`,
         MD_projection_fibres= `PC_res_MD_projection_fibres`) %>%
  select(f.eid, MD_global, MD_association_fibres, MD_thalamic_radiations, MD_projection_fibres)


## Merge with full dataframe 
UKB_imaging_covariates <- left_join(UKB_imaging_covariates, UKB_imaging_covariates_temp, by = "f.eid")


## Calculate general FA measures
## ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
## Association/Commissural fibres


## Association/Commissural fibres

# Cingulate gyrus part of cingulum (left) - f.25492
# Cingulate gyrus part of cingulum (right) - f.25493
# Inferior fronto-occipital fasciculus (left) - f.25500
# Inferior fronto-occipital fasciculus (right) - f.25501
# Inferior longitudinal fasciculus (left) - f.25502
# Inferior longitudinal fasciculus (right) - f.25503
# Parahippocampal part of cingulum (left) - f.25494
# Parahippocampal part of cingulum (right) - f.25495
# Superior longitudinal fasciculus (left) - f.25509
# Superior longitudinal fasciculus (right) - f.25510
# Uncinate fasciculus (left) - f.25513
# Uncinate fasciculus (right) - f.25514
# Forceps major - f.25498
# Forceps minor - f.25499


FA_association_fibres <- c("f.25492.2.0", "f.25493.2.0", "f.25500.2.0", "f.25501.2.0", "f.25502.2.0", "f.25503.2.0",
                           "f.25494.2.0", "f.25495.2.0", "f.25509.2.0", "f.25510.2.0", "f.25513.2.0", "f.25514.2.0",
                           "f.25498.2.0", "f.25499.2.0")

## Thalamic radiations

# Anterior thalamic radiation (left)  - f.25490
# Anterior thalamic radiation (right) - f.25491
# Posterior thalamic radiation (left) - f.25507
# Posterior thalamic radiation (right) - f.25508
# Superior thalamic radiation (left) - f.25511
# Superior thalamic radiation (right) - f.25512

FA_thalamic_radiations <- c("f.25490.2.0", "f.25491.2.0", "f.25507.2.0", "f.25508.2.0", "f.25511.2.0", "f.25512.2.0")


## Projection fibres

# Acousticc radiation (left) - f.25488
# Acousticc radiation (right) - f.25489
# Corticospinal tract (left) - f.25496
# Corticospinal tract (right) - f.25497
# Medial lemniscus (left) - f.25505
# Medial lemniscus (right) - f.25506
# Middle cerebellar peduncle - f.25504

FA_projection_fibres <- c("f.25488.2.0", "f.25489.2.0", "f.25496.2.0", "f.25497.2.0", "f.25505.2.0", "f.25506.2.0", "f.25504.2.0")

## FA_global
FA_global <- c(FA_association_fibres, FA_thalamic_radiations, FA_projection_fibres)

## Remove participants without DTI information and store in temporary file (PCA does not allow NAs)
UKB_imaging_covariates_temp <- UKB_imaging_covariates[!is.na(UKB_imaging_covariates$f.25516.2.0),]

## Run PCA in each general FA measure
PC_res_FA_association_fibres <- princomp((UKB_imaging_covariates_temp[,FA_association_fibres]),scores = TRUE)$scores[,1]

PC_res_FA_association_fibres_PCA <- princomp(UKB_imaging_covariates_temp[,FA_association_fibres])
temp <- PC_res_FA_association_fibres_PCA$sdev^2 / sum(PC_res_FA_association_fibres_PCA$sdev^2)
PC_res_FA_association_fibres_VE <- temp[1]
                                             
PC_res_FA_thalamic_radiations <- princomp(UKB_imaging_covariates_temp[,FA_thalamic_radiations],scores = TRUE)$scores[,1]

PC_res_FA_thalamic_radiations_PCA <- princomp(UKB_imaging_covariates_temp[,FA_thalamic_radiations])
temp <- PC_res_FA_thalamic_radiations_PCA$sdev^2 / sum(PC_res_FA_thalamic_radiations_PCA$sdev^2)
PC_res_FA_thalamic_radiations_VE <- temp[1]

PC_res_FA_projection_fibres <- princomp(UKB_imaging_covariates_temp[,FA_projection_fibres],scores = TRUE)$scores[,1]

PC_res_FA_projection_fibres_PCA <- princomp(UKB_imaging_covariates_temp[,FA_projection_fibres])
temp <- PC_res_FA_projection_fibres_PCA$sdev^2 / sum(PC_res_FA_projection_fibres_PCA$sdev^2)
PC_res_FA_projection_fibres_VE <- temp[1]

PC_res_FA_global <- princomp(UKB_imaging_covariates_temp[,FA_global],scores = TRUE)$scores[,1]

PC_res_FA_global_PCA <- princomp(UKB_imaging_covariates_temp[,FA_global])
temp <- PC_res_FA_global_PCA$sdev^2 / sum(PC_res_FA_global_PCA$sdev^2)
PC_res_FA_global_VE <- temp[1]

## Add 1st PC to temporary dataframe
UKB_imaging_covariates_temp <- cbind(as.data.frame(UKB_imaging_covariates_temp$f.eid), PC_res_FA_association_fibres, PC_res_FA_thalamic_radiations, PC_res_FA_projection_fibres, PC_res_FA_global)

## Rename column containing 1st PC of each general FA meaure and keep only ID and general FA meaures
UKB_imaging_covariates_temp <- UKB_imaging_covariates_temp %>%
  rename(f.eid = "UKB_imaging_covariates_temp$f.eid",
         FA_global= "PC_res_FA_global",
         FA_association_fibres= "PC_res_FA_association_fibres",
         FA_thalamic_radiations= "PC_res_FA_thalamic_radiations",
         FA_projection_fibres= "PC_res_FA_projection_fibres") %>%
  select(f.eid, FA_global, FA_association_fibres, FA_thalamic_radiations, FA_projection_fibres)


## Merge with full dataframe 
UKB_imaging_covariates <- left_join(UKB_imaging_covariates, UKB_imaging_covariates_temp, by = "f.eid")

## Save variance explained

DTI_VE_df <- data.frame(
  "DTI_measure" = c("MD global", "MD association fibres", "MD projection fibres", "MD thalamic associations", "FA global", "FA association fibres", "FA projection fibres", "FA thalamic associations"),
  "variance_explained" = c(PC_res_MD_global_VE, PC_res_MD_association_fibres_VE, PC_res_MD_projection_fibres_VE, PC_res_MD_thalamic_radiations_VE, PC_res_FA_global_VE, PC_res_FA_association_fibres_VE, PC_res_FA_projection_fibres_VE, PC_res_FA_thalamic_radiations_VE)
)

write.csv(DTI_VE_df, "/output/DTI_PCA_variance_explained.csv", row.names = F)

```

## Get descriptive names for all imaging variables

```{r Convert FIDs to descriptive names}

## Load in UKB DKW field ID key file
UKB_cortical_region_key <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/resources/UKB_imaging_keys/UKB_cortical_region_field_IDs.csv", header = FALSE)
UKB_cortical_region_key <- UKB_cortical_region_key %>%
  rename(feild_ID = V1,
         cortical_region_measure = V2)

## Load in UKB FA field ID key file
UKB_FA_tract_key <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/resources/UKB_imaging_keys/UKB_FA_field_IDs.csv", header = FALSE)
UKB_FA_tract_key <- UKB_FA_tract_key %>%
  rename(feild_ID = V1,
         FA_value = V2)


## Load in UKB FA field ID key file
UKB_MD_tract_key <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/resources/UKB_imaging_keys/UKB_MD_field_IDs.csv", header = FALSE)
UKB_MD_tract_key <- UKB_MD_tract_key %>%
  rename(feild_ID = V1,
         MD_value = V2)


## Load in UKB FIRST subcortical field ID key file
UKB_subcortical_region_key <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/resources/UKB_imaging_keys/UKB_subcortical_field_IDs.csv", header = FALSE)
UKB_subcortical_region_key <- UKB_subcortical_region_key %>%
  rename(feild_ID = V1,
         subcortical_volume = V2)


## Change names of cortical volumes from FIDs to descriptive names
## get names of individual cortical volumes
cortical_volume_FIDs <- paste0("f.", c(26789:26821, 26890:26922), ".2.0")
for (i in 1:length(cortical_volume_FIDs)){
  ## Get name of individual cortical volume from key file to save mean to dataframe
  cortical_volume <- UKB_cortical_region_key$cortical_region_measure[grepl(substring(cortical_volume_FIDs[i], 3,7), UKB_cortical_region_key$feild_ID)]
  cortical_volume <- sub(("Volume of "), "volume_", cortical_volume) ## simplify prefix indicating measurement type
  cortical_volume <- gsub(" \\(", "_", cortical_volume) ## replace " (" with underscore
  cortical_volume <- gsub(" hemisphere)", "", cortical_volume) ## remove " hemisphere"
  
  ## rename column
  names(UKB_imaging_covariates)[names(UKB_imaging_covariates) == cortical_volume_FIDs[i]] <- cortical_volume
}

## Change names of cortical areas from FIDs to descriptive names
## get names of individual cortical areas
cortical_area_FIDs <- paste0("f.", c(26722:26754, 26823:26855), ".2.0")

for (i in 1:length(cortical_area_FIDs)){
  ## Get name of individual cortical area from key file to save mean to dataframe
  cortical_area <- UKB_cortical_region_key$cortical_region_measure[grepl(substring(cortical_area_FIDs[i], 3,7), UKB_cortical_region_key$feild_ID)]
  cortical_area <- sub(("Area of "), "area_", cortical_area) ## simplify prefix indicating measurement type
  cortical_area <- gsub(" \\(", "_", cortical_area) ## replace " (" with underscore
  cortical_area <- gsub(" hemisphere)", "", cortical_area) ## remove " hemisphere"
  
  ## rename column
  names(UKB_imaging_covariates)[names(UKB_imaging_covariates) == cortical_area_FIDs[i]] <- cortical_area
}

## Change names of cortical thickness from FIDs to descriptive names
## get names of individual cortical thicknesss
cortical_thickness_FIDs <- paste0("f.", c(26756:26788, 26857:26889), ".2.0")

for (i in 1:length(cortical_thickness_FIDs)){
  ## Get name of individual cortical thickness from key file to save mean to dataframe
  cortical_thickness <- UKB_cortical_region_key$cortical_region_measure[grepl(substring(cortical_thickness_FIDs[i], 3,7), UKB_cortical_region_key$feild_ID)]
  cortical_thickness <- sub(("Mean thickness of "), "thickness_", cortical_thickness) ## simplify prefix indicating measurement type
  cortical_thickness <- gsub(" \\(", "_", cortical_thickness) ## replace " (" with underscore
  cortical_thickness <- gsub(" hemisphere)", "", cortical_thickness) ## remove " hemisphere"
  
  ## rename column
  names(UKB_imaging_covariates)[names(UKB_imaging_covariates) == cortical_thickness_FIDs[i]] <- cortical_thickness
}

## Change names of MD tracts from FIDs to descriptive names
## get names of individual bilateral MD tracts
MD_bilateral_FIDs <- paste0("f.", c(25515:25524, 25527:25530, 25532:25541), ".2.0")

for (i in 1:length(MD_bilateral_FIDs)){
  
  ## Get name of individual bilateral MD tracts from key file to save mean to dataframe
  MD_bilateral_tract_name <- UKB_MD_tract_key$MD_value[grepl(substring(MD_bilateral_FIDs[i], 3,7), UKB_MD_tract_key$feild_ID)]
  MD_bilateral_tract_name <- sub(("Weighted-mean MD in tract"), "MD", MD_bilateral_tract_name) ## simplify prefix indicating measurement type
  MD_bilateral_tract_name <- trimws(MD_bilateral_tract_name) ## trim leading and trailing whitespace
  MD_bilateral_tract_name <- gsub(" ", "_", MD_bilateral_tract_name) ## replace any spaces with underscores
  MD_bilateral_tract_name <- gsub("\\(", "", MD_bilateral_tract_name) ## remove brackets
  MD_bilateral_tract_name <- gsub(")", "", MD_bilateral_tract_name) ## remove brackets
  
  ## rename column
  names(UKB_imaging_covariates)[names(UKB_imaging_covariates) == MD_bilateral_FIDs[i]] <- MD_bilateral_tract_name
}

## get names of individual unilateral MD tracts
MD_unilateral_FIDs <- paste0("f.", c(25525, 25526, 25531), ".2.0")

for (i in 1:length(MD_unilateral_FIDs)){
  
  ## Get name of individual bilateral MD tracts from key file to save mean to dataframe
  MD_bilateral_tract_name <- UKB_MD_tract_key$MD_value[grepl(substring(MD_unilateral_FIDs[i], 3,7), UKB_MD_tract_key$feild_ID)]
  MD_bilateral_tract_name <- sub(("Weighted-mean MD in tract"), "MD", MD_bilateral_tract_name) ## simplify prefix indicating measurement type
  MD_bilateral_tract_name <- trimws(MD_bilateral_tract_name) ## trim leading and trailing whitespace
  MD_bilateral_tract_name <- gsub(" ", "_", MD_bilateral_tract_name) ## replace any spaces with underscores
  MD_bilateral_tract_name <- gsub("\\(", "", MD_bilateral_tract_name) ## remove brackets
  MD_bilateral_tract_name <- gsub(")", "", MD_bilateral_tract_name) ## remove brackets
  
  ## rename column
  names(UKB_imaging_covariates)[names(UKB_imaging_covariates) == MD_unilateral_FIDs[i]] <- MD_bilateral_tract_name
}


## Change names of FA tracts from FIDs to descriptive names
## get names of individual bilateral FA tracts
FA_bilateral_FIDs <- paste0("f.", c(25488:25497, 25500:25503, 25505:25514), ".2.0")

for (i in 1:length(FA_bilateral_FIDs)){
  
  ## Get name of individual bilateral FA tracts from key file to save mean to dataframe
  FA_bilateral_tract_name <- UKB_FA_tract_key$FA_value[grepl(substring(FA_bilateral_FIDs[i], 3,7), UKB_FA_tract_key$feild_ID)]
  FA_bilateral_tract_name <- sub(("Weighted-mean FA in tract"), "FA", FA_bilateral_tract_name) ## simplify prefix indicating measurement type
  FA_bilateral_tract_name <- trimws(FA_bilateral_tract_name) ## trim leading and trailing whitespace
  FA_bilateral_tract_name <- gsub(" ", "_", FA_bilateral_tract_name) ## replace any spaces with underscores
  FA_bilateral_tract_name <- gsub("\\(", "", FA_bilateral_tract_name) ## remove brackets
  FA_bilateral_tract_name <- gsub(")", "", FA_bilateral_tract_name) ## remove brackets
  
  ## rename column
  names(UKB_imaging_covariates)[names(UKB_imaging_covariates) == FA_bilateral_FIDs[i]] <- FA_bilateral_tract_name
}

## get names of individual unilateral FA tracts
FA_unilateral_FIDs <- paste0("f.", c(25498, 25499, 25504), ".2.0")

for (i in 1:length(FA_unilateral_FIDs)){
  
  ## Get name of individual bilateral FA tracts from key file to save mean to dataframe
  FA_bilateral_tract_name <- UKB_FA_tract_key$FA_value[grepl(substring(FA_unilateral_FIDs[i], 3,7), UKB_FA_tract_key$feild_ID)]
  FA_bilateral_tract_name <- sub(("Weighted-mean FA in tract"), "FA", FA_bilateral_tract_name) ## simplify prefix indicating measurement type
  FA_bilateral_tract_name <- trimws(FA_bilateral_tract_name) ## trim leading and trailing whitespace
  FA_bilateral_tract_name <- gsub(" ", "_", FA_bilateral_tract_name) ## replace any spaces with underscores
  FA_bilateral_tract_name <- gsub("\\(", "", FA_bilateral_tract_name) ## remove brackets
  FA_bilateral_tract_name <- gsub(")", "", FA_bilateral_tract_name) ## remove brackets
  
  ## rename column
  names(UKB_imaging_covariates)[names(UKB_imaging_covariates) == FA_unilateral_FIDs[i]] <- FA_bilateral_tract_name
}

## Subcortical Volumes
sub_cortical_volume_FIDs <- paste0("f.", c(25011:25024), ".2.0")

for (i in 1:length(sub_cortical_volume_FIDs)){
  
  ## Get name of individual bilateral FA tracts from key file to save mean to dataframe
  subcortical_volume <- UKB_subcortical_region_key$subcortical_volume[grepl(substring(sub_cortical_volume_FIDs[i], 3,7), UKB_subcortical_region_key$feild_ID)]
  subcortical_volume <- sub(("Volume of "), "sub_cortical_volume_", subcortical_volume) ## simplify prefix indicating measurement type
  subcortical_volume <- trimws(subcortical_volume) ## trim leading and trailing whitespace
  subcortical_volume <- gsub(" ", "_", subcortical_volume) ## replace any spaces with underscores
  subcortical_volume <- gsub("\\(", "", subcortical_volume) ## remove brackets
  subcortical_volume <- gsub(")", "", subcortical_volume) ## remove brackets
  
  ## rename column
  names(UKB_imaging_covariates)[names(UKB_imaging_covariates) == sub_cortical_volume_FIDs[i]] <- subcortical_volume
}
```

## Remove outliers - remove any imaging structre value with SD <3 from mean

```{r Remove outliers}
DTI_brain_structure_names <- grep("FA|MD", names(UKB_imaging_covariates), value = TRUE)
T1_brain_structure_names <- grep("volume|area|thickness", names(UKB_imaging_covariates), value = TRUE)

## Get a dataframe only containing brain structures for outlier removal - will be used to calculate zscores
DTI_brain_structure_df <- UKB_imaging_covariates[,DTI_brain_structure_names]
T1_brain_structure_df <- UKB_imaging_covariates[,T1_brain_structure_names]

## Calculate Z score for each imaging measure
DTI_z_scores <- as.data.frame(sapply(DTI_brain_structure_df, function(DTI_brain_structure_df) (abs(DTI_brain_structure_df-mean(DTI_brain_structure_df, na.rm = TRUE))/sd(DTI_brain_structure_df, na.rm = TRUE))))    

T1_z_scores <- as.data.frame(sapply(T1_brain_structure_df, function(T1_brain_structure_df) (abs(T1_brain_structure_df-mean(T1_brain_structure_df, na.rm = TRUE))/sd(T1_brain_structure_df, na.rm = TRUE))))    

## Get ID of those with at least one brain structure > 5SD from mean
DTI_outlier_ids <- unique(UKB_imaging_covariates$f.eid[rowSums(DTI_z_scores>5, na.rm = T) > 0])
T1_outlier_ids <- unique(UKB_imaging_covariates$f.eid[rowSums(T1_z_scores>5, na.rm = T) > 0])

outlier_ids <- unique(c(DTI_outlier_ids, T1_outlier_ids))

## Save list of outliers
write.csv(as.data.frame(DTI_outlier_ids), "/resources/DTI_imaging_outliers.csv", quote = F)
write.csv(as.data.frame(T1_outlier_ids), "/resources/T1_imaging_outliers.csv", quote = F)
write.csv(as.data.frame(outlier_ids), "/resources/imaging_outliers.csv", quote = F)
```
## Scale all brain structures and continuous covariates

```{r Scale data }
## Scale brain strucutre data
UKB_imaging_covariates$age_raw <- UKB_imaging_covariates$age
UKB_imaging_covariates$BMI_raw <- UKB_imaging_covariates$BMI

UKB_imaging_covariates[,c(all_brain_structure_names, "age", "age_squared", "BMI", "ICV", "X_position", "Y_position", "Z_position")] <- as.data.frame(scale(UKB_imaging_covariates[,c(all_brain_structure_names, "age", "age_squared", "BMI", "ICV", "X_position", "Y_position", "Z_position")]))

```

## Create long-formatted dataframe with bilateral data (needed for LME)

```{r Long-formatting}

# Function to reshape wide imaging format into long imaging format
bilateral_imag <- function(imagnames,hemnames,idname,n,data,covariates){
  
  # Function arguments:
  # Imagnames: names to be assigned to new time-varying variables (i.e. for left and right anterior thalamic radiation, this will be "anterior thalamic radiation")
  # Hemnames: names of left and right hemisphere brain regions to loop through in order to reshape into long format
  # ID: participant ID name (this differs between cohorts)
  # N: number of bilateral regions (i.e. if 12 lh and 12 rh, N=24)
  # Data: dataset in which the variables are
  # Covariates: dataset with covariates to merge after reshaping the data
  
  long_format <- lapply(seq(1,n,2), function(x)
    melt(setDT(data), id = idname, 
         measure = c(hemnames[x],hemnames[x + 1]), 
         value.name = c(imagnames[x]),
         variable.name = "hemi"))
  
  # Reduce all lists to the first column (participant ID), last column (the 2 hemispheres put together), and "hemi"
  imag_only = lapply(long_format, function(x) x %>% select(c(1,2,3)))
  
  # Cbind all imag variables
  imag_cbind = list.cbind(imag_only)
  
  # Remove duplicated columns - this will leave "hemi" with level names that corresponded to the first imaging measure (this is a residual from the first list element - when you remove duplicates of a column, only the first obs remains); this should be turned into "1" and "2"
  imag_cbind_deduped=imag_cbind %>% subset(., select = which(!duplicated(names(.))))
  
  # Merge with the other covs
  imag_cbind_deduped=merge(imag_cbind_deduped,covariates,by=idname)
  
  # Turn "hemi" to factor with level "1" and "2"
  levels(imag_cbind_deduped$hemi) <- c(1,2)
  
  # return
  return(imag_cbind_deduped)
}

## Get list of names to be assigned to new time-varying variables
imaging_structures_bilateral_imagnames <- grep("_right", names(UKB_imaging_covariates), value = TRUE)
imaging_structures_bilateral_imagnames <- gsub("_right", "", imaging_structures_bilateral_imagnames)

## Get list of left and right hemisphere brain regions to loop through in order to reshape into long format
imaging_structures_bilateral_hemnames <- c(paste0(imaging_structures_bilateral_imagnames, "_right"), paste0(imaging_structures_bilateral_imagnames, "_left"))

## replicate all elements in array containing brain strucutre name
imaging_structures_bilateral_imagnames<- rep(imaging_structures_bilateral_imagnames,2)

## Sort lists alphebetically
imaging_structures_bilateral_imagnames <- sort(imaging_structures_bilateral_imagnames)
imaging_structures_bilateral_hemnames<- sort(imaging_structures_bilateral_hemnames)

## Subset UKB_imaging_covariates to get covariates + independent variables (inflam biomarkers) + IDs
UKB_imaging_just_covariates <- UKB_imaging_covariates %>%
  select(f.eid, sex, age, age_squared, BMI, assessment_centre_first_imaging, ICV, X_position, Y_position, Z_position)

## Run wide to long format function on all bilateral structures
UKB_imaging_covariates_long <- bilateral_imag(imagnames = imaging_structures_bilateral_imagnames,
                                hemnames = imaging_structures_bilateral_hemnames,
                                idname = "f.eid",
                                n = length(imaging_structures_bilateral_hemnames),
                                data = UKB_imaging_covariates,
                                covariates =  UKB_imaging_just_covariates)

## Check data had been processed correctly
head(UKB_imaging_covariates_long)
dim(UKB_imaging_covariates)
dim(UKB_imaging_covariates_long)
```

## Save short and long-formatted preprocessed brain structure data

```{r Save to project folder}
write.csv(UKB_imaging_covariates, "/resources/UKB_imaging_covariates_qc.csv", quote = F)
write.csv(UKB_imaging_covariates_long, "/resources/UKB_imaging_covariates_long_qc.csv", quote = F)
```
