==> 6_UKB_CP_MDD_imaging_present.Rmd <==
---
title: "Association of chronic pain and depression with brain structure"
author: "Hannah Casey"
date: "6/17/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,fig.width=12, fig.height=12)
# Read in relevant R packages
library(tidyverse)
library(grid)
library(kableExtra)
library(plotly)
library(wesanderson)
library(ggrepel)
library(ggforce)
library(ggseg)
library(patchwork)
library(RColorBrewer)
library(flextable)
library(ggpubr)
#devtools::install_github("teunbrand/ggh4x")
library(ggh4x)
# Include width and height for figures in the knitr::opts_chunk$set code above
```

**Pre-processing and QC - brain structure data:**

- Outliers removed (>5 SD)
- Global DTI measures calculated using PCA

----------------------------------------------------------------------------------------------------------------------------
Models:

- GLM for unilatearal brain phenotypes
- LME for bilateral brain phenotypes where measures of hemisphere were treated as repeated measures
  - No hemishphere interactions on association between DNAm score and bilateral phenotype observed
  
Covariates: 
- Sex
- Age
- Age^2
- UKB imaging assessment centre
- ICV (only for measures of brain volume)

Multiple comparision correction: FDR
  - n = number of brain structure phenotypes per imaging subcategory
  
#### Get results

```{r Load in and format data, echo=FALSE,  fig.height=50}
# Read in results
res_CP_MDD_brain_structure_all <- read.csv("/Users/hannahcasey/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_assoc_all.csv",header=T)
res_CP_MDD_brain_structure_all$sample <- "All\n"

res_CP_MDD_structure_sex_interaction_all <- read.csv("/Users/hannahcasey/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_sex_interaction_all.csv",header=T)
res_CP_MDD_structure_sex_interaction_all$sample <- "Sex Interaction\n(Female - Male)"
res_CP_MDD_structure_sex_interaction_all$p.adjust <- res_CP_MDD_structure_sex_interaction_all$p.value
res_CP_MDD_structure_sex_interaction_all <- res_CP_MDD_structure_sex_interaction_all %>%
  select(brain_phenotype, brain_structure_type, disorder, beta, p.adjust, sample)

res_CP_MDD_brain_structure_female <- read.csv("/Users/hannahcasey/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_assoc_female.csv",header=T)
res_CP_MDD_brain_structure_female$sample <- "Female\n"

res_CP_MDD_brain_structure_male <- read.csv("/Users/hannahcasey/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_assoc_male.csv",header=T)
res_CP_MDD_brain_structure_male$sample <- "Male\n"


res_CP_MDD_structure_disorder_interaction_all <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_disorder_interaction_all.csv")
res_CP_MDD_structure_disorder_interaction_all$sample <- "All\n"

res_CP_MDD_structure_disorder_sex_interaction_all <- read.csv("/Users/hannahcasey/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_disorder_sex_interaction_all.csv",header=T)
res_CP_MDD_structure_disorder_sex_interaction_all$sample <- "Sex Interaction\n(Female - Male)"
res_CP_MDD_structure_disorder_sex_interaction_all$p.adjust <- res_CP_MDD_structure_disorder_sex_interaction_all$p.value
res_CP_MDD_structure_disorder_sex_interaction_all <- res_CP_MDD_structure_disorder_sex_interaction_all %>%
  select(brain_phenotype, brain_structure_type, disorder, beta, p.adjust, sample)

res_CP_MDD_structure_disorder_interaction_male <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_disorder_interaction_male.csv")
res_CP_MDD_structure_disorder_interaction_male$sample <- "Male\n"

res_CP_MDD_structure_disorder_interaction_female <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_disorder_interaction_female.csv")
res_CP_MDD_structure_disorder_interaction_female$sample <- "Female\n"


## Combine results into single dataframes
res_CP_MDD_brain_structure <- rbind(res_CP_MDD_brain_structure_all, res_CP_MDD_brain_structure_female, res_CP_MDD_brain_structure_male)
res_CP_MDD_brain_structure <- res_CP_MDD_brain_structure %>%
  select(brain_phenotype, brain_structure_type, disorder, beta, p.adjust, sample) %>%
  rbind(res_CP_MDD_structure_sex_interaction_all)

res_CP_MDD_structure_disorder_interaction <- rbind(res_CP_MDD_structure_disorder_interaction_all, res_CP_MDD_structure_disorder_interaction_male, res_CP_MDD_structure_disorder_interaction_female)
res_CP_MDD_structure_disorder_interaction <- res_CP_MDD_structure_disorder_interaction %>%
  select(brain_phenotype, brain_structure_type, disorder, beta, p.adjust, sample) %>%
  rbind(res_CP_MDD_structure_disorder_sex_interaction_all)

## Create new columns to reformat brain phenotypes for plotting
res_CP_MDD_brain_structure$brain_phenotype_plot <-  gsub("MD", "", res_CP_MDD_brain_structure$brain_phenotype)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("FA", "", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("volume", "", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("area", "", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("thickness", "", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("sub_cortical_", "", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("inferior", "inferior ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("superior", "superior ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("anterior", "anterior ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("caudal", "caudal ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("middle", "middle ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("lateral", "lateral ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("isthmus", "isthmus ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("posterior", "posterior ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("rostral", "rostral ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("pars", "pars ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("medial", "medial ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("\\s+", " ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("_|[.]", " ", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("fronto ", "fronto-", res_CP_MDD_brain_structure$brain_phenotype_plot)


res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <-  gsub("MD", "", res_CP_MDD_structure_disorder_interaction$brain_phenotype)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("FA", "", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("volume", "", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("area", "", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("thickness", "", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("sub_cortical_", "", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("inferior", "inferior ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("superior", "superior ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("anterior", "anterior ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("caudal", "caudal ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("middle", "middle ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("lateral", "lateral ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("isthmus", "isthmus ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("posterior", "posterior ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("rostral", "rostral ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("pars", "pars ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("medial", "medial ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("\\s+", " ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("_|[.]", " ", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)
res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot <- gsub("fronto ", "fronto-", res_CP_MDD_structure_disorder_interaction$brain_phenotype_plot)

## Reorder brain phenotype and subgroups for plotting
res_CP_MDD_brain_structure$brain_structure_type  <- ordered(res_CP_MDD_brain_structure$brain_structure_type , levels = c("Regional Subcortical Volume", "Regional Cortical Volume","Global and Lobal Cortical Volume", "Regional Cortical Area", "Global and Lobal Cortical Area", "Regional Cortical Thickness","Global and Lobal Cortical Thickness", "Fractional Anisotropy", "General Fractional Anisotropy","Mean Diffusivity", "General Mean Diffusivity"))

res_CP_MDD_structure_disorder_interaction$brain_structure_type  <- ordered(res_CP_MDD_structure_disorder_interaction$brain_structure_type , levels = c("Regional Subcortical Volume", "Regional Cortical Volume","Global and Lobal Cortical Volume", "Regional Cortical Area", "Global and Lobal Cortical Area", "Regional Cortical Thickness","Global and Lobal Cortical Thickness", "Fractional Anisotropy", "General Fractional Anisotropy","Mean Diffusivity", "General Mean Diffusivity"))

# Include yes/no column to determine shape based on p-value
res_CP_MDD_brain_structure$`P-value (adjusted/not adjusted*) < 0.01` = ifelse(res_CP_MDD_brain_structure$p.adjust < 0.05, "Yes","No")
res_CP_MDD_structure_disorder_interaction$significant <- ifelse(res_CP_MDD_structure_disorder_interaction$p.adjust < 0.05,"Yes" , "No")

## Add column indicating if results are positive or negative (will be used in p-value graph later)
res_CP_MDD_brain_structure$Direction <- ifelse(sign(res_CP_MDD_brain_structure$beta) < 0,"Negative" , "Positive")
res_CP_MDD_structure_disorder_interaction$Direction <- ifelse(sign(res_CP_MDD_structure_disorder_interaction$beta) < 0,"Negative" , "Positive")
```

#### Sample demographics

```{r, echo=FALSE, results="asis", comment=NA}
## Load in file with summary demographics
demographics <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/descriptives.csv", header = T)

flextable(demographics) %>%
  autofit() %>%
  fit_to_width(7) 
```

#### Main Results - Brain structures {.tabset}

#### Depression association analysis
```{r, echo=FALSE, results="asis", comment=NA}
## Transform p.adjsut to use as y-axis
res_CP_MDD_brain_structure$p.adjust.minus.log <- -log10(res_CP_MDD_brain_structure$p.adjust)

## Recode brain strucutre subtypes with abbreviations
brain_structure_abbreviations <- c(
  "Regional Subcortical Volume" = "SCV",
  "Regional Cortical Volume" = "RCV",
  "Global and Lobal Cortical Volume" = "GCV",
  "Regional Cortical Area" = "RCA",
  "Global and Lobal Cortical Area" = "GCA",
  "Regional Cortical Thickness" = "RCT",
  "Global and Lobal Cortical Thickness" = "GCT",
  "Fractional Anisotropy" = "FA",
  "General Fractional Anisotropy" = "GFA",
  "Mean Diffusivity" = "MD",
  "General Mean Diffusivity" = "GMD"
)

## Get unique brain structure types
brain_structure_types <- levels(res_CP_MDD_brain_structure$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)

# Define custom colors
custom_colors <- c("Positive" = "#0072B2", "Negative" = "#D55E00", "Not Significant" = "grey60")
threshold <- -log10(0.05)

## Depression full sample----

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

## Define jitter settings
set.seed(1997)  # Ensure reproducibility
jitter_width <- 0.2  # Adjust as needed

## Merge shading data with the main dataset
depression_all_df <- res_CP_MDD_brain_structure %>%
  filter(disorder == "depression" & sample == "All\n") %>%
  mutate(
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
    ),
    brain_structure_abbreviation = recode(brain_structure_type, !!!brain_structure_abbreviations)
  )
  
## Identify the top significant finding in each brain structure category
top_findings <- depression_all_df %>%
  filter(p.adjust < 0.05) %>%
  group_by(brain_structure_type) %>%
  arrange(p.adjust, desc(abs(beta))) %>%  # First by p-value, then by largest absolute beta
  slice(1)


depression_all_plot <- ggplot(
  depression_all_df,
  aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  ## Add threshold reference line
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Add labels only for the most significant point in each brain structure category
  geom_text_repel(
    data = top_findings,
    aes(label = brain_phenotype_plot),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    segment.size = 0.3,
    color = "black",
    alpha = 0.8,
    max.overlaps = Inf,
    nudge_y = 0.5,
    force = 10,
    show.legend = FALSE
  ) +
  ## Axis formatting
  scale_x_continuous(
    breaks = sort(unique(depression_all_df$brain_structure_numeric)),
    labels = brain_structure_abbreviations,  
  ) +
  scale_y_continuous(limits = c(0, 6)) +
  ## Custom Theme
  theme_classic() +
  theme(
    text = element_text(size = 17),
    axis.text.x = element_text(size = 17),
    legend.position = "bottom"
  ) +
  ## Custom Colors
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
  guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ## Labels and Title
  xlab("") + 
  ylab("-log10(p)") +
  ggtitle('Full Sample')

## Depression sex interaction ----

## Merge shading data with the main dataset
depression_sex_interaction_df <- res_CP_MDD_brain_structure %>%
  filter(disorder == "depression" & sample == "Sex Interaction\n(Female - Male)") %>%
  group_by(brain_structure_type) %>%
  filter(any(p.adjust < 0.05)) %>% ## Remove brain structure groups without any significant findings
  ungroup() %>%
  mutate(brain_structure_type = factor(brain_structure_type, levels = levels(res_CP_MDD_brain_structure$brain_structure_type))) %>% ## Set order
  arrange(brain_structure_type) %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

## Get unique brain structure types
brain_structure_types <- unique(depression_sex_interaction_df$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)
brain_structure_abbreviations_filtered <- brain_structure_abbreviations[names(brain_structure_abbreviations) %in% brain_structure_types]

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

depression_sex_interaction_plot <- ggplot(
  depression_sex_interaction_df,
  aes(x = jittered_x, y = p.adjust.minus.log, fill = Significance, colour = Significance, shape = Direction, stroke = 1)) +
  ## Shading
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Labels
  geom_text_repel(
    aes(label = ifelse(p.adjust < 0.05, as.character(brain_phenotype_plot), '')),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    color = 'black',
    segment.size = 0.3,
    alpha = 0.7,
    max.overlaps = Inf,
    nudge_y = 0.3,
    direction = "both",
    min.segment.length = 0.5,
    force = 10,
    force_pull = 0,
    show.legend = FALSE
  ) +
    scale_x_continuous(
    breaks = sort(unique(depression_sex_interaction_df$brain_structure_numeric)),  # Keep original positions
    labels = brain_structure_abbreviations_filtered  # Show corresponding names
  ) +
    scale_y_continuous(
    limits =  c(0, 6)
  ) +
  theme_classic() +
  theme(
    text = element_text(size=17),
    axis.text.x = element_text(size=17), 
    legend.position = "none"  # Remove legend
  ) +
  xlab("") + 
  ylab("-log10(p)") +
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ggtitle('Sex Interaction (Female - Male)')

## Depression male sample ----

## Merge shading data with the main dataset
depression_male_df <- res_CP_MDD_brain_structure %>%
  filter(disorder == "depression" & (sample == "Sex Interaction\n(Female - Male)" | sample == "Male\n")) %>%
  group_by(brain_structure_type) %>%
  filter(any(p.adjust < 0.05)) %>% ## Remove brain structure groups not included in interaction analysis
  ungroup() %>%
  filter(disorder == "depression" & sample == "Male\n") %>%
  mutate(brain_structure_type = factor(brain_structure_type, levels = levels(res_CP_MDD_brain_structure$brain_structure_type))) %>% ## Set order
  arrange(brain_structure_type) %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

## Get unique brain structure types
brain_structure_types <- unique(depression_male_df$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)
brain_structure_abbreviations_filtered <- brain_structure_abbreviations[names(brain_structure_abbreviations) %in% brain_structure_types]

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

depression_male_plot <- ggplot(
  depression_male_df,
  aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Labels
  geom_text_repel(
    aes(label = ifelse(p.adjust < 0.05, as.character(brain_phenotype_plot), '')),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    color = 'black',
    segment.size = 0.3,
    alpha = 0.7,
    max.overlaps = Inf,
    nudge_y = 0.3,
    direction = "both",
    min.segment.length = 0.5,
    force = 10,
    force_pull = 0,
    show.legend = FALSE
  ) +
    scale_x_continuous(
    breaks = sort(unique(depression_male_df$brain_structure_numeric)),  # Keep original positions
    labels = brain_structure_abbreviations_filtered  # Show corresponding names
  ) +
    scale_y_continuous(
    limits =  c(0, 6)
  ) +
  theme_classic() +
  theme(
    text = element_text(size=17),
    axis.text.x = element_text(size=17), 
    legend.position = "none"  # Remove legend
  ) +
  xlab("") + 
  ylab("-log10(p)") +
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ggtitle('Male Sample')

## Depression female sample ----

## Merge shading data with the main dataset
depression_female_df <- res_CP_MDD_brain_structure %>%
  filter(disorder == "depression" & (sample == "Sex Interaction\n(Female - Male)" | sample == "Female\n")) %>%
  group_by(brain_structure_type) %>%
  filter(any(p.adjust < 0.05)) %>% ## Remove brain structure groups without any significant findings
  ungroup() %>%
  filter(disorder == "depression" & sample == "Female\n") %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

## Get unique brain structure types
brain_structure_types <- unique(depression_female_df$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)
brain_structure_abbreviations_filtered <- brain_structure_abbreviations[names(brain_structure_abbreviations) %in% brain_structure_types]

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

depression_female_plot <- ggplot(
  depression_female_df,
  aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  # Labels
  geom_text_repel(
    aes(label = ifelse(p.adjust < 0.05, as.character(brain_phenotype_plot), '')),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    color = 'black',
    segment.size = 0.3,
    alpha = 0.7,
    max.overlaps = Inf,
    nudge_y = 0.3,
    direction = "both",
    min.segment.length = 0.5,
    force = 10,
    force_pull = 0,
    show.legend = FALSE
  ) +
    scale_x_continuous(
    breaks = sort(unique(depression_female_df$brain_structure_numeric)),  # Keep original positions
    labels = brain_structure_abbreviations_filtered  # Show corresponding names
  ) +
  scale_y_continuous(
    limits =  c(0, 6)
  ) +
  theme_classic() +
  theme(
    text = element_text(size=17),
    axis.text.x = element_text(size=17),
    legend.position = "none"  # Remove legend
  ) +
  xlab("") + 
  ylab("-log10(p)") +
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ggtitle('Female Sample')

tiff("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/depression.tiff", units="in", width=15, height=15, res=300)

depression_all_plot / (depression_sex_interaction_plot + depression_female_plot + depression_male_plot) +
   plot_layout(heights = unit(c(20, 1), c('cm', 'null')))

dev.off()
```

####  Chronic Pain association analysis
```{r, echo=FALSE, results="asis", comment=NA}
## Transform p.adjsut to use as y-axis
res_CP_MDD_brain_structure$p.adjust.minus.log <- -log10(res_CP_MDD_brain_structure$p.adjust)

## Get unique brain structure types
brain_structure_types <- levels(res_CP_MDD_brain_structure$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)

## Chronic Pain full sample----

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

## Define jitter settings
set.seed(1997)  # Ensure reproducibility
jitter_width <- 0.2  # Adjust as needed

## Merge shading data with the main dataset
chronic_pain_all_df <- res_CP_MDD_brain_structure %>%
  filter(disorder == "chronic_pain_status" & sample == "All\n") %>%
  mutate(brain_structure_type = factor(brain_structure_type, levels = levels(res_CP_MDD_brain_structure$brain_structure_type))) %>% ## Set order
  arrange(brain_structure_type) %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

## Identify the top significant finding in each brain structure category
top_findings <- chronic_pain_all_df %>%
  filter(p.adjust < 0.05) %>%
  group_by(brain_structure_type) %>%
  arrange(p.adjust, desc(abs(beta))) %>%  # First by p-value, then by largest absolute beta
  slice(1)


chronic_pain_all_plot <- ggplot(
  chronic_pain_all_df,
    aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  ## Add threshold reference line
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Add labels only for the most significant point in each brain structure category
  geom_text_repel(
    data = top_findings,
    aes(label = brain_phenotype_plot),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    segment.size = 0.3,
    color = "black",
    alpha = 0.8,
    max.overlaps = Inf,
    nudge_y = 0.5,
    force = 10,
    show.legend = FALSE
  ) +
  ## Axis formatting
  scale_x_continuous(
    breaks = sort(unique(chronic_pain_all_df$brain_structure_numeric)),  # Keep original positions
    labels = brain_structure_abbreviations # Show corresponding names
  ) +
  scale_y_continuous(limits = c(0, 17)) +
  ## Custom Theme
  theme_classic() +
  theme(
    text = element_text(size = 17),
    axis.text.x = element_text(size = 17),
    legend.position = "bottom"
  ) +
  ## Custom Colors
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ## Labels and Title
  xlab("") + 
  ylab("-log10(p)") +
  ggtitle('Full Sample')

## Chronic Pain sex interaction ----

## Merge shading data with the main dataset
chronic_pain_sex_interaction_df <- res_CP_MDD_brain_structure %>%
  filter(disorder == "chronic_pain_status" & sample == "Sex Interaction\n(Female - Male)") %>%
  group_by(brain_structure_type) %>%
  filter(any(p.adjust < 0.05)) %>% ## Remove brain structure groups without any significant findings
  ungroup() %>%
  mutate(brain_structure_type = factor(brain_structure_type, levels = levels(res_CP_MDD_brain_structure$brain_structure_type))) %>% ## Set order
  arrange(brain_structure_type) %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

## Get unique brain structure types
brain_structure_types <- unique(chronic_pain_sex_interaction_df$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)
brain_structure_abbreviations_filtered <- brain_structure_abbreviations[names(brain_structure_abbreviations) %in% brain_structure_types]

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

chronic_pain_sex_interaction_plot <- ggplot(
  chronic_pain_sex_interaction_df,
  aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Labels
  geom_text_repel(
    aes(label = ifelse(p.adjust < 0.05, as.character(brain_phenotype_plot), '')),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    color = 'black',
    segment.size = 0.3,
    alpha = 0.7,
    max.overlaps = Inf,
    nudge_y = 0.3,
    direction = "both",
    min.segment.length = 0.5,
    force = 10,
    force_pull = 0,
    show.legend = FALSE
  ) +
    scale_x_continuous(
    breaks = sort(unique(chronic_pain_sex_interaction_df$brain_structure_numeric)),  # Keep original positions
    labels = brain_structure_abbreviations_filtered  # Show corresponding names
  ) +
    scale_y_continuous(
    limits = c(0, 17)
  ) +
  theme_classic() +
  theme(
    text = element_text(size=17),
    axis.text.x = element_text(size=17), 
    legend.position = "none"  # Remove legend
  ) +
  xlab("") + 
  ylab("-log10(p)") +
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ggtitle('Sex Interaction (Female - Male)')

## Chronic Pain male sample ----

## Merge shading data with the main dataset
chronic_pain_male_df <- res_CP_MDD_brain_structure %>%
  filter(disorder == "chronic_pain_status" & (sample == "Sex Interaction\n(Female - Male)" | sample == "Male\n")) %>%
  group_by(brain_structure_type) %>%
  filter(any(p.adjust < 0.05)) %>% ## Remove brain structure groups not included in interaction analysis
  ungroup() %>%
  filter(disorder == "chronic_pain_status" & sample == "Male\n") %>%
  mutate(brain_structure_type = factor(brain_structure_type, levels = levels(res_CP_MDD_brain_structure$brain_structure_type))) %>% ## Set order
  arrange(brain_structure_type) %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

## Get unique brain structure types
brain_structure_types <- unique(chronic_pain_male_df$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)
brain_structure_abbreviations_filtered <- brain_structure_abbreviations[names(brain_structure_abbreviations) %in% brain_structure_types]

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

chronic_pain_male_plot <- ggplot(
  chronic_pain_male_df,
  aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Labels
  geom_text_repel(
    aes(label = ifelse(p.adjust < 0.05, as.character(brain_phenotype_plot), '')),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    color = 'black',
    segment.size = 0.3,
    alpha = 0.7,
    max.overlaps = Inf,
    nudge_y = 0.3,
    direction = "both",
    min.segment.length = 0.5,
    force = 10,
    force_pull = 0,
    show.legend = FALSE
  ) +
    scale_x_continuous(
    breaks = sort(unique(chronic_pain_male_df$brain_structure_numeric)),  # Keep original positions
    labels = brain_structure_abbreviations_filtered  # Show corresponding names
  ) +
    scale_y_continuous(
    limits = c(0, 17)
  ) +
  theme_classic() +
  theme(
    text = element_text(size=17),
    axis.text.x = element_text(size=17), 
    legend.position = "none"  # Remove legend
  ) +
  xlab("") + 
  ylab("-log10(p)") +
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ggtitle('Male Sample')

## Chronic Pain female sample ----

## Merge shading data with the main dataset
chronic_pain_female_df <- res_CP_MDD_brain_structure %>%
  filter(disorder == "chronic_pain_status" & (sample == "Sex Interaction\n(Female - Male)" | sample == "Female\n")) %>%
  group_by(brain_structure_type) %>%
  filter(any(p.adjust < 0.05)) %>% ## Remove brain structure groups without any significant findings
  ungroup() %>%
  filter(disorder == "chronic_pain_status" & sample == "Female\n") %>%
  mutate(brain_structure_type = factor(brain_structure_type, levels = levels(res_CP_MDD_brain_structure$brain_structure_type))) %>% ## Set order
  arrange(brain_structure_type) %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

## Get unique brain structure types
brain_structure_types <- unique(chronic_pain_female_df$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)
brain_structure_abbreviations_filtered <- brain_structure_abbreviations[names(brain_structure_abbreviations) %in% brain_structure_types]

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

chronic_pain_female_plot <- ggplot(
  chronic_pain_female_df,
  aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Labels
  geom_text_repel(
    aes(label = ifelse(p.adjust < 0.05, as.character(brain_phenotype_plot), '')),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    color = 'black',
    segment.size = 0.3,
    alpha = 0.7,
    max.overlaps = Inf,
    nudge_y = 0.3,
    direction = "both",
    min.segment.length = 0.5,
    force = 10,
    force_pull = 0,
    show.legend = FALSE
  ) +
    scale_x_continuous(
    breaks = sort(unique(chronic_pain_female_df$brain_structure_numeric)),  # Keep original positions
    labels = brain_structure_abbreviations_filtered  # Show corresponding names
  ) +
  scale_y_continuous(
    limits = c(0, 17)
  ) +
  theme_classic() +
  theme(
    text = element_text(size=17),
    axis.text.x = element_text(size=17),
    legend.position = "none"  # Remove legend
  ) +
  xlab("") + 
  ylab("-log10(p)") +
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ggtitle('Female Sample')

tiff("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/chronic_pain.tiff", units="in", width=15, height=15, res=300)

chronic_pain_all_plot / (chronic_pain_sex_interaction_plot + chronic_pain_female_plot + chronic_pain_male_plot) +
   plot_layout(heights = unit(c(22, 1), c('cm', 'null')))

dev.off()
```

####  Disorder interaction analysis
```{r, echo=FALSE, results="asis", comment=NA}
## Transform p.adjsut to use as y-axis
res_CP_MDD_structure_disorder_interaction$p.adjust.minus.log <- -log10(res_CP_MDD_structure_disorder_interaction$p.adjust)

## Get unique brain structure types
brain_structure_types <- levels(res_CP_MDD_structure_disorder_interaction$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)

## Disorder Interaction full sample----

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

## Define jitter settings
set.seed(1997)  # Ensure reproducibility
jitter_width <- 0.2  # Adjust as needed

## Merge shading data with the main dataset
disorder_interaction_all_df <- res_CP_MDD_structure_disorder_interaction %>%
  filter(sample == "All\n") %>%
  mutate(brain_structure_type = factor(brain_structure_type, levels = levels(res_CP_MDD_structure_disorder_interaction$brain_structure_type))) %>% ## Set order
  arrange(brain_structure_type) %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

disorder_interaction_all_plot <- ggplot(
  disorder_interaction_all_df,
  aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Labels
  geom_text_repel(
    aes(label = ifelse(p.adjust < 0.05, as.character(brain_phenotype_plot), '')),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    color = 'black',
    segment.size = 0.3,
    alpha = 0.7,
    max.overlaps = Inf,
    nudge_y = 0.3,
    direction = "both",
    min.segment.length = 0.5,
    force = 10,
    force_pull = 0,
    show.legend = FALSE
  ) +
    scale_x_continuous(
    breaks = sort(unique(disorder_interaction_all_df$brain_structure_numeric)) # Keep original positions
    #labels = depression_all_df$brain_structure_type # Show corresponding names
    ) +
  scale_y_continuous(
    limits = c(0, 3.5)
  ) +
  theme_classic() +
  theme(
    text = element_text(size=17),
    axis.text.x = element_text(size=17), 
    legend.position = "none"  # Remove legend
  ) +
  xlab("") + 
  ylab("-log10(p)") +
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ggtitle('Full Sample')

## Disorder-sex interaction ----

## Merge shading data with the main dataset
disorder_sex_interaction_df <- res_CP_MDD_structure_disorder_interaction %>%
  filter(sample == "Sex Interaction\n(Female - Male)") %>%
  group_by(brain_structure_type) %>%
  filter(any(p.adjust < 0.05)) %>% ## Remove brain structure groups without any significant findings
  ungroup() %>%
  mutate(brain_structure_type = factor(brain_structure_type, levels = levels(res_CP_MDD_structure_disorder_interaction$brain_structure_type))) %>% ## Set order
  arrange(brain_structure_type) %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

## Get unique brain structure types
brain_structure_types <- unique(disorder_sex_interaction_df$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)
brain_structure_abbreviations_filtered <- brain_structure_abbreviations[names(brain_structure_abbreviations) %in% brain_structure_types]

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

disorder_sex_interaction_plot <- ggplot(
  disorder_sex_interaction_df,
  aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Labels
  geom_text_repel(
    aes(label = ifelse(p.adjust < 0.05, as.character(brain_phenotype_plot), '')),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    color = 'black',
    segment.size = 0.3,
    alpha = 0.7,
    max.overlaps = Inf,
    nudge_y = 0.3,
    direction = "both",
    min.segment.length = 0.5,
    force = 10,
    force_pull = 0,
    show.legend = FALSE
  ) +
    scale_x_continuous(
    breaks = sort(unique(disorder_sex_interaction_df$brain_structure_numeric)),  # Keep original positions
    labels = brain_structure_abbreviations_filtered  # Show corresponding names
  ) +
    scale_y_continuous(
    limits = c(0, 3.5)
  ) +
  theme_classic() +
  theme(
    text = element_text(size=17),
    axis.text.x = element_text(size=17), 
    legend.position = "none"  # Remove legend
  ) +
  xlab("") + 
  ylab("-log10(p)") +
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ggtitle('Sex Interaction (Female - Male)')

## Disorder interaction male sample ----

## Merge shading data with the main dataset
disorder_interaction_male_df <- res_CP_MDD_structure_disorder_interaction %>%
  filter(sample == "Sex Interaction\n(Female - Male)" | sample == "Male\n") %>%
  group_by(brain_structure_type) %>%
  filter(any(p.adjust < 0.05)) %>% ## Remove brain structure groups not included in interaction analysis
  ungroup() %>%
  filter(sample == "Male\n") %>%
  mutate(brain_structure_type = factor(brain_structure_type, levels = levels(res_CP_MDD_structure_disorder_interaction$brain_structure_type))) %>% ## Set order
  arrange(brain_structure_type) %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

## Get unique brain structure types
brain_structure_types <- unique(disorder_interaction_male_df$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)
brain_structure_abbreviations_filtered <- brain_structure_abbreviations[names(brain_structure_abbreviations) %in% brain_structure_types]

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

disorder_interaction_male_plot <- ggplot(
  disorder_interaction_male_df,
  aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Labels
  geom_text_repel(
    aes(label = ifelse(p.adjust < 0.05, as.character(brain_phenotype_plot), '')),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    color = 'black',
    segment.size = 0.3,
    alpha = 0.7,
    max.overlaps = Inf,
    nudge_y = 0.3,
    direction = "both",
    min.segment.length = 0.5,
    force = 10,
    force_pull = 0,
    show.legend = FALSE
  ) +
    scale_x_continuous(
    breaks = sort(unique(disorder_interaction_male_df$brain_structure_numeric)),  # Keep original positions
    labels = brain_structure_abbreviations_filtered  # Show corresponding names
  ) +
    scale_y_continuous(
    limits = c(0, 3.5)
  ) +
  theme_classic() +
  theme(
    text = element_text(size=17),
    axis.text.x = element_text(size=17), 
    legend.position = "none"  # Remove legend
  ) +
  xlab("") + 
  ylab("-log10(p)") +
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ggtitle('Male Sample')

## Disorder interaction female sample ----

## Merge shading data with the main dataset
disorder_interaction_female_df <- res_CP_MDD_structure_disorder_interaction %>%
  filter(sample == "Sex Interaction\n(Female - Male)" | sample == "Female\n") %>%
  group_by(brain_structure_type) %>%
  filter(any(p.adjust < 0.05)) %>% ## Remove brain structure groups not included in interaction analysis
  ungroup() %>%
  filter(sample == "Female\n") %>%
  mutate(brain_structure_type = factor(brain_structure_type, levels = levels(res_CP_MDD_structure_disorder_interaction$brain_structure_type))) %>% ## Set order
  arrange(brain_structure_type) %>%
  mutate( ## Apply jitter manually to x-axis (brain structure type)
    brain_structure_numeric = as.numeric(factor(brain_structure_type)),  # Convert to numeric
    jittered_x = brain_structure_numeric + runif(n(), -jitter_width, jitter_width),  # Apply jitter
    Significance = case_when(
      p.adjust < 0.05 & Direction == "Positive" ~ "Positive",
      p.adjust < 0.05 & Direction == "Negative" ~ "Negative",
      TRUE ~ "Not Significant"
      )
  )

## Get unique brain structure types
brain_structure_types <- unique(disorder_interaction_female_df$brain_structure_type)
brain_structure_positions <- seq_along(brain_structure_types)
brain_structure_abbreviations_filtered <- brain_structure_abbreviations[names(brain_structure_abbreviations) %in% brain_structure_types]

## Create shading data for background
shading_all_df <- data.frame(
  xmin = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] - 0.5,
  xmax = brain_structure_positions[seq(1, length(brain_structure_positions), by = 2)] + 0.5,
  ymin = -Inf,
  ymax = Inf
)

disorder_interaction_female_plot <- ggplot(
  disorder_interaction_female_df,
  aes(x = jittered_x, y = p.adjust.minus.log, shape = Direction, stroke = 1)
) +
  ## Background shading for alternate structure types
  geom_rect(data = shading_all_df, 
            aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), 
            fill = "lightgrey", alpha = 0.2, inherit.aes = FALSE) +
  scale_shape_manual(values = c(25, 24)) + 
  ## Color coding significant positive/negative points
  geom_point(aes(color = Significance, fill = Significance, alpha = 0.8), 
             size = 3, stroke = 1) +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "red", size = 1, alpha = 0.4) +
  ## Labels
  geom_text_repel(
    aes(label = ifelse(p.adjust < 0.05, as.character(brain_phenotype_plot), '')),
    size = 7,
    box.padding = 0.5,
    point.padding = 0.5,
    segment.color = 'black',
    color = 'black',
    segment.size = 0.3,
    alpha = 0.7,
    max.overlaps = Inf,
    nudge_y = 0.3,
    direction = "both",
    min.segment.length = 0.5,
    force = 10,
    force_pull = 0,
    show.legend = FALSE
  ) +
    scale_x_continuous(
    breaks = sort(unique(disorder_interaction_female_df$brain_structure_numeric)),  # Keep original positions
    labels = brain_structure_abbreviations_filtered  # Show corresponding names
  ) +
    scale_y_continuous(
    limits = c(0, 3.5)
  ) +
  theme_classic() +
  theme(
    text = element_text(size=17),
    axis.text.x = element_text(size=17), 
    legend.position = "none"  # Remove legend
  ) +
  xlab("") + 
  ylab("-log10(p)") +
  scale_fill_manual(values = custom_colors) +  
  scale_color_manual(values = custom_colors) +  
    guides(colour = "none", fill = "none", shape = "none", alpha = "none") + 
  ggtitle('Female Sample')

tiff("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/disorder_interaction.tiff", units="in", width=15, height=10, res=300)
(disorder_sex_interaction_plot + disorder_interaction_female_plot + disorder_interaction_male_plot)
dev.off()
```

#### Common alterations in chronic pain and depression
```{r, echo=FALSE, results="asis", comment=NA}
# Filter significant results
depression_sig <- depression_all_df %>% filter(p.adjust < 0.05)
chronic_pain_sig <- chronic_pain_all_df %>% filter(p.adjust < 0.05)

## Number of associations in common
length(intersect(depression_sig$brain_phenotype, chronic_pain_sig$brain_phenotype))

## Proportion of chronic pain in depression
sum(depression_sig$brain_phenotype %in% chronic_pain_sig$brain_phenotype)/length(depression_sig$brain_phenotype)

## Proportion of depression in chronic pain
sum(chronic_pain_sig$brain_phenotype %in% depression_sig$brain_phenotype)/length(chronic_pain_sig$brain_phenotype)

depression_sig$brain_phenotype <- gsub("_right|_left", "", depression_sig$brain_phenotype)
chronic_pain_sig$brain_phenotype <- gsub("_right|_left", "", chronic_pain_sig$brain_phenotype)

# Merge on common brain phenotype
common_sig <- inner_join(depression_sig, chronic_pain_sig, by = "brain_phenotype", suffix = c("_dep", "_pain"))

# Filter where the direction of effect is the same (both beta values positive or both negative)
common_sig_same_direction <- common_sig %>%
  filter(sign(beta_dep) == sign(beta_pain))

```

### Volume in chronic pain
```{r, echo=FALSE, results="asis", comment=NA}

CP_sign_volume <- chronic_pain_all_df %>%
  filter(p.adjust < 0.05) %>%
  filter(brain_structure_type == "Regional Cortical Volume") %>%
  select(brain_phenotype_plot) %>%
  mutate(brain_phenotype_plot = gsub(" left| right", "", brain_phenotype_plot)) %>%
  unique()


CP_sign_area <- chronic_pain_all_df %>%
  filter(p.adjust < 0.05) %>%
  filter(brain_structure_type == "Regional Cortical Area") %>%
  select(brain_phenotype_plot) %>%
  mutate(brain_phenotype_plot = gsub(" left| right", "", brain_phenotype_plot)) %>%
  unique()


CP_sign_thickness <- chronic_pain_all_df %>%
  filter(p.adjust < 0.05) %>%
  filter(Direction == "Negative") %>%
  filter(brain_structure_type == "Regional Cortical Thickness") %>%
  select(brain_phenotype_plot) %>%
  mutate(brain_phenotype_plot = gsub(" left| right", "", brain_phenotype_plot)) %>%
  unique()


## Lower volume with lower area and thickness
length(CP_sign_volume$brain_phenotype_plot)
sum(CP_sign_volume$brain_phenotype_plot %in% CP_sign_area$brain_phenotype_plot)
sum(CP_sign_volume$brain_phenotype_plot %in% CP_sign_thickness$brain_phenotype_plot)
sum(CP_sign_volume$brain_phenotype_plot %in% CP_sign_thickness$brain_phenotype_plot & CP_sign_volume$brain_phenotype_plot %in% CP_sign_area$brain_phenotype_plot)


## Lower volumes withput lower area and thickness
CP_sign_volume$brain_phenotype_plot[!CP_sign_volume$brain_phenotype_plot %in% CP_sign_area$brain_phenotype_plot & !CP_sign_volume$brain_phenotype_plot %in% CP_sign_thickness$brain_phenotype_plot]


CP_sign_area$brain_phenotype_plot[CP_sign_area$brain_phenotype_plot %in% CP_sign_volume$brain_phenotype_plot]

CP_sign_thickness$brain_phenotype_plot[CP_sign_thickness$brain_phenotype_plot %in% CP_sign_volume$brain_phenotype_plot]


brain_results <- chronic_pain_all_df %>%
  mutate(
    base_region = str_remove(brain_phenotype, "^(left_|right_)"),
    base_region = str_remove(base_region, "(volume|area|thickness)$")
  )

significant <- brain_results %>%
  filter(p.adjust < 0.05)

CP_region_morph_counts <- significant %>%
  count(brain_phenotype_plot) %>%
  filter(n >= 2)

```
