==> 8_UKB_brain_structure_CP_MDD_supplementary.Rmd <==
---
title: "Structural brain correlates of current chronic pain and depression: cross-sectional study in UK Biobank Supplementary Materials"
output: 
  word_document:
    reference_docx: ~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/writeup/RMD_template.docx
---

```{r setup, include=FALSE}
library(kableExtra)
library(dplyr)
library(flextable)
library(stringr)
library(tidyr)
```

# Table of contents:

## 1. Supplementary Methods

### 1.1. Neurological disorders excluded

### 1.2. Brain structures full details

## 2. Supplementary Results

### 2.1. Full results from main analysis

\newpage

# 1. Supplementary Methods

\n

\n

## 1.1. Neurological disorders excluded

Neurological disorder status was determined based on self-reported illnesses recorded in UK Biobank data fields 20001 and 20002.

```{r Neurological, echo=F}

UKB_neurological <- data.frame(Disorder = c("Benign neuroma", "Brain abscess/intracranial abscess", "Brain haemorrhage", "Cerebral aneurysm", "Cerebral palsy", "Chronic/degenerative neurological problem", "dementia/alzheimers/cognitive impairment", "Encephalitis", "Epilepsy", "Fracture skull/head", "Head injury", "Ischaemic stroke", "Meningitis", "Motor Neurone Disease", "Multiple Sclerosis", "Nervous system infection", "Neurological injury/trauma", "Other demyelinating disease (not Multiple Sclerosis)", "Other neurological problem", "Parkinson’s Disease", "Spina Bifida", "Stroke", "Subarachnoid haemorrhage", "Subdural haemorrhage/haematoma", "Transient ischaemic attack", "Meningioma/benign meningeal tumour"), `Coding` = c("1683", 1245, 1491, 1425, 1433, 1258, 1263, 1246, 1264, 1626, 1266, 1583, 1247, 1259, 1261, 1244, 1240, 1397, 1434, 1262, 1524, 1081, 1086, 1083, 1082, 1659))

UKB_neurological_cancer <- data.frame(Disorder = c("Brain cancer/primary malignant brain tumour", "Meningeal cancer/malignant meningioma"), `Coding` = c("1032", 1031))

flextable(UKB_neurological) %>%
  autofit() %>%
  fit_to_width(7) %>%
  theme_booktabs()
  
```

Table S1 Non-cancer neurological conditions excluded for based on data field 20002.

\newpage

```{r Neurological cancer, echo=F}
flextable(UKB_neurological_cancer) %>%
  autofit() %>%
  fit_to_width(7)
```

Table S2 Cancer neurological conditions excluded for based on data field 20001.

\newpage

## 1.2. Brain structures full details

Measures of cortical volume, thickness and surface area in UK Biobank

```{r Cortical measures, echo=F, warning=F}
frontal_lobe <- c("superior frontal gyrus" , "rostral middle frontal", "caudal middle frontal", "pars orbitalis", "pars triangularis", "pars opercularis", "frontal pole", "lateral orbitofrontal", "precentral gyrus", "paracentral cortex")
temporal_lobe <- c("insula", "superior temporal", "transverse temporal", "banks STS", "middle temporal gyrus" , "iferior temporal gyrus" , "fusiform" , "parahippocampal" , "entorhinal")
parietal_lobe <- c("postcentral gyrus", "paracentral cortex", "superior parietal cortex", "inferior parietal cortex", "supramarginal gyrus", "precuneus")
occipital_lobe <- c("lateral occipital cortex", "cuneus" , "pericalcarine cortex" , "lingual gyrus")
cingulate_lobe <- c("rostral ACC", "caudal ACC", "posterior cingulate cortex", "cingulate isthmus")

cortical_measures <- data_frame(`Cortical Region` = c(frontal_lobe, temporal_lobe, parietal_lobe, occipital_lobe, cingulate_lobe), `Lobe` = c(rep("frontal", length(frontal_lobe)), rep("temporal", length(temporal_lobe)), rep("parietal", length(parietal_lobe)), rep("occipital", length(occipital_lobe)), rep("cingulate", length(cingulate_lobe))))

flextable(cortical_measures) %>%
  autofit() %>%
  fit_to_width(7) 
```

Table S3 Cortical regions in UK Biobank and cortical lobes derived from "Desikan-Killiany-Tourville" atlas

\newpage

```{r subcortical measures, echo=F,  warning=F}

subcortical_volumes <- data_frame(`Subcortical Region` = c("thalamus", "caudate", "putamen", "pallidum", "hippocampus", "amygdala", "accumbens"))

flextable(subcortical_volumes) %>%
  autofit() %>%
  fit_to_width(7) 
```

Table S4 Regional Subcortical Volumes in UK Biobank derived using FIRST (FMRIB's Integrated Registration and Segmentation Tool)

\newpage

```{r WM measures, echo=F,  warning=F}

association_fibres <- c("cingulate gyrus part of cingulum" , "inferior fronto-occipital fasciculus", "inferior longitudinal fasciculus", "parahippocampal part of cingulum", "superior longitudinal fasciculus", "uncinate fasciculus", "forceps major", "forceps minor")

thalamic_radiations <- c("anterior thalamic radiation", "posterior thalamic radiation", "superior thalamic radiation")

projection_fibres <- c("acoustic radiation", "corticospinal tract", "medial lemniscus" , "middle cerebellar peduncle")


WM_measures <- data_frame(`White Matter Tract` = c(association_fibres, thalamic_radiations, projection_fibres), `General Measure` = c(rep("association fibre", length(association_fibres)), rep("thalamic radiations", length(thalamic_radiations)), rep("projection fibres", length(projection_fibres))))

flextable(WM_measures) %>%
  autofit() %>%
  fit_to_width(7) 

```

Table S5 White matter tracts in UK Biobank and general measures derived using AutoPtx.

\newpage

```{r PCA DTI, echo=F,  warning=F}

PCA_DTI_VE <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/DTI_PCA_variance_explained.csv")

PCA_DTI_VE$variance_explained <- round(PCA_DTI_VE$variance_explained, digits = 3)

PCA_DTI_VE <- PCA_DTI_VE %>%
  rename("Variance Explained" = variance_explained,
         "White Matter General Measures" = DTI_measure)

flextable(PCA_DTI_VE) %>%
  autofit() %>%
  fit_to_width(7) 


```
Table S6 Variance explained by the first principal component analysis (PCA) component for each global and general measure of white matter tracts.

\newpage

## 2. Supplementary Results

\n

\n

### 2.1. Full results from analysis

```{r All results, echo=F, warning=F}
# Read in results 
res_CP_MDD_brain_structure_all <- read.csv("/Users/hannahcasey/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_assoc_all.csv",header=T)
res_CP_MDD_brain_structure_all$sample <- "All"

res_CP_MDD_structure_sex_interaction_all <- read.csv("/Users/hannahcasey/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_sex_interaction_all.csv",header=T)
res_CP_MDD_structure_sex_interaction_all$sample <- "Sex Interaction (Females - Males)"
res_CP_MDD_structure_sex_interaction_all$p.adjust <- res_CP_MDD_structure_sex_interaction_all$p.value
res_CP_MDD_structure_sex_interaction_all <- res_CP_MDD_structure_sex_interaction_all %>%
  select(brain_phenotype, brain_structure_type, disorder, beta, p.adjust, sample)

res_CP_MDD_brain_structure_female <- read.csv("/Users/hannahcasey/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_assoc_female.csv",header=T)
res_CP_MDD_brain_structure_female$sample <- "Female"

res_CP_MDD_brain_structure_male <- read.csv("/Users/hannahcasey/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_assoc_male.csv",header=T)
res_CP_MDD_brain_structure_male$sample <- "Male"

res_disorder_interaction_brain_structure_all <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_disorder_interaction_all.csv")
res_disorder_interaction_brain_structure_all$sample <- "All"

res_disorder_sex_interaction_brain_structure_all <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_disorder_sex_interaction_all.csv")
res_disorder_sex_interaction_brain_structure_all$sample <- "Sex Interaction (Females - Males)"
res_disorder_sex_interaction_brain_structure_all$p.adjust <- res_disorder_sex_interaction_brain_structure_all$p.value
res_disorder_sex_interaction_brain_structure_all <- res_disorder_sex_interaction_brain_structure_all %>%
  select(brain_phenotype, brain_structure_type, beta, p.adjust, sample)

res_disorder_interaction_brain_structure_female <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_disorder_interaction_female.csv")
res_disorder_interaction_brain_structure_female$sample <- "Female"

res_disorder_interaction_brain_structure_male <- read.csv("~/Desktop/PhD/projects/UKB_CP_MDD_brain_structure/output/CP_MDD_structure_disorder_interaction_male.csv")
res_disorder_interaction_brain_structure_male$sample <- "Male"

## Combine results into single dataframes
res_CP_MDD_brain_structure <- rbind(res_CP_MDD_brain_structure_all, res_CP_MDD_brain_structure_female, res_CP_MDD_brain_structure_male)
res_CP_MDD_brain_structure <- res_CP_MDD_brain_structure %>%
  select(brain_phenotype, brain_structure_type, disorder, beta, p.adjust, sample) %>%
  rbind(res_CP_MDD_structure_sex_interaction_all)

res_disorder_interaction_brain_structure <- rbind(res_disorder_interaction_brain_structure_all, res_disorder_interaction_brain_structure_female, res_disorder_interaction_brain_structure_male)
res_disorder_interaction_brain_structure <- res_disorder_interaction_brain_structure %>%
  select(brain_phenotype, brain_structure_type, beta, p.adjust, sample) %>%
  rbind(res_disorder_sex_interaction_brain_structure_all)

## Create new columns to reformat brain phenotypes for plotting
res_CP_MDD_brain_structure$brain_phenotype_plot <-  gsub("MD", "MD in", res_CP_MDD_brain_structure$brain_phenotype)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("FA", "FA in", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("volume", "Volume of", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("area", "Area of", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("thickness", "Thickness of", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("sub_cortical_Volume", "Regional Subcortical Volume", res_CP_MDD_brain_structure$brain_phenotype_plot)
res_CP_MDD_brain_structure$brain_phenotype_plot <- gsub("_|[.]", " ", res_CP_MDD_brain_structure$brain_phenotype_plot)

res_disorder_interaction_brain_structure$brain_phenotype_plot <-  gsub("MD", "MD in", res_disorder_interaction_brain_structure$brain_phenotype)
res_disorder_interaction_brain_structure$brain_phenotype_plot <- gsub("FA", "FA in", res_disorder_interaction_brain_structure$brain_phenotype_plot)
res_disorder_interaction_brain_structure$brain_phenotype_plot <- gsub("volume", "Volume of", res_disorder_interaction_brain_structure$brain_phenotype_plot)
res_disorder_interaction_brain_structure$brain_phenotype_plot <- gsub("area", "Area of", res_disorder_interaction_brain_structure$brain_phenotype_plot)
res_disorder_interaction_brain_structure$brain_phenotype_plot <- gsub("thickness", "Thickness of", res_disorder_interaction_brain_structure$brain_phenotype_plot)
res_disorder_interaction_brain_structure$brain_phenotype_plot <- gsub("sub_cortical_Volume", "Regional Subcortical Volume", res_disorder_interaction_brain_structure$brain_phenotype_plot)
res_disorder_interaction_brain_structure$brain_phenotype_plot <- gsub("\\.[1-9]", "", res_disorder_interaction_brain_structure$brain_phenotype_plot)
res_disorder_interaction_brain_structure$brain_phenotype_plot <-  gsub("_|[.]", " ", res_disorder_interaction_brain_structure$brain_phenotype_plot)

## Reorder brain phenotype and subgroups for plotting
res_CP_MDD_brain_structure$brain_structure_type  <- ordered(res_CP_MDD_brain_structure$brain_structure_type , levels = c("Regional Subcortical Volume", "Regional Cortical Volume","Global and Lobal Cortical Volume", "Regional Cortical Area", "Global and Lobal Cortical Area", "Regional Cortical Thickness","Global and Lobal Cortical Thickness", "Fractional Anisotropy", "Mean Diffusivity","General Fractional Anisotropy", "General Mean Diffusivity"))

res_disorder_interaction_brain_structure$brain_structure_type  <- ordered(res_disorder_interaction_brain_structure$brain_structure_type , levels = c("Regional Subcortical Volume", "Regional Cortical Volume","Global and Lobal Cortical Volume", "Regional Cortical Area", "Global and Lobal Cortical Area", "Regional Cortical Thickness","Global and Lobal Cortical Thickness", "Fractional Anisotropy", "Mean Diffusivity","General Fractional Anisotropy", "General Mean Diffusivity"))

res_CP_MDD_brain_structure_CP <- res_CP_MDD_brain_structure[res_CP_MDD_brain_structure$disorder == "chronic_pain_status",]
res_CP_MDD_brain_structure_MDD <- res_CP_MDD_brain_structure[res_CP_MDD_brain_structure$disorder == "depression",]


## Chronic pain
reshaped <- reshape(res_CP_MDD_brain_structure_CP, idvar = c("brain_phenotype_plot", "brain_structure_type"), timevar = "sample", direction = "wide")
reshaped <- reshaped[grepl("brain_phenotype_plot|beta|p.adjust|brain_structure_type",names(reshaped))]

## Add astriks to significant betas
reshaped$beta.All <- round(reshaped$beta.All, 3)
temp <- reshaped$beta.All
reshaped$beta.All = ifelse(reshaped$p.adjust.All < 0.05, paste0(temp, "*"), reshaped$beta.All)
reshaped$beta.All = ifelse(reshaped$p.adjust.All < 0.01, paste0(temp, "**"), reshaped$beta.All)
reshaped$beta.All = ifelse(reshaped$p.adjust.All < 0.001, paste0(temp, "***"), reshaped$beta.All)

reshaped$`beta.Sex Interaction (Females - Males)` <- round(reshaped$`beta.Sex Interaction (Females - Males)`, 3)
temp <- reshaped$`beta.Sex Interaction (Females - Males)`
reshaped$`beta.Sex Interaction (Females - Males)` = ifelse(reshaped$`p.adjust.Sex Interaction (Females - Males)` < 0.05, paste0(temp, "*"), reshaped$`beta.Sex Interaction (Females - Males)`)
reshaped$`beta.Sex Interaction (Females - Males)` = ifelse(reshaped$`p.adjust.Sex Interaction (Females - Males)` < 0.01, paste0(temp, "**"), reshaped$`beta.Sex Interaction (Females - Males)`)
reshaped$`beta.Sex Interaction (Females - Males)` = ifelse(reshaped$`p.adjust.Sex Interaction (Females - Males)` < 0.001, paste0(temp, "***"), reshaped$`beta.Sex Interaction (Females - Males)`)

reshaped$beta.Female <- round(reshaped$beta.Female, 3)
temp <- reshaped$beta.Female
reshaped$beta.Female = ifelse(reshaped$p.adjust.Female < 0.05, paste0(temp, "*"), reshaped$beta.Female)
reshaped$beta.Female = ifelse(reshaped$p.adjust.Female < 0.01, paste0(temp, "**"), reshaped$beta.Female)
reshaped$beta.Female = ifelse(reshaped$p.adjust.Female < 0.001, paste0(temp, "***"), reshaped$beta.Female)

reshaped$beta.Male <- round(reshaped$beta.Male, 3)
temp <- reshaped$beta.Male
reshaped$beta.Male = ifelse(reshaped$p.adjust.Male < 0.05, paste0(temp, "*"), reshaped$beta.Male)
reshaped$beta.Male = ifelse(reshaped$p.adjust.Male < 0.01, paste0(temp, "**"), reshaped$beta.Male)
reshaped$beta.Male = ifelse(reshaped$p.adjust.Male < 0.001, paste0(temp, "***"), reshaped$beta.Male)

cat_order <- c("Regional Subcortical Volume", "Regional Cortical Volume","Global and Lobal Cortical Volume", "Regional Cortical Area", "Global and Lobal Cortical Area", "Regional Cortical Thickness","Global and Lobal Cortical Thickness", "Fractional Anisotropy", "Mean Diffusivity","General Fractional Anisotropy", "General Mean Diffusivity")


reshaped <- reshaped %>% 
  arrange(brain_phenotype_plot) %>%
  arrange(factor(brain_structure_type, levels = cat_order))

res_CP <- reshaped %>%
  rename(`Brain Strucutre Phenotype` = brain_phenotype_plot,
         `Full Sample Beta` = beta.All,
         `Sex Interaction (Females - Males)` = `beta.Sex Interaction (Females - Males)`,
         `Female Sample Beta` = beta.Female,
         `Male Sample Beta` = beta.Male) %>%
  select(`Brain Strucutre Phenotype`, `Full Sample Beta`, `Sex Interaction (Females - Males)`, `Female Sample Beta`, `Male Sample Beta`)

## Depression
reshaped <- reshape(res_CP_MDD_brain_structure_MDD, idvar = c("brain_phenotype_plot", "brain_structure_type"), timevar = "sample", direction = "wide")
reshaped <- reshaped[grepl("brain_phenotype_plot|beta|p.adjust|brain_structure_type",names(reshaped))]

## Add astriks to significant betas
reshaped$beta.All <- round(reshaped$beta.All, 3)
temp <- reshaped$beta.All
reshaped$beta.All = ifelse(reshaped$p.adjust.All < 0.05, paste0(temp, "*"), reshaped$beta.All)
reshaped$beta.All = ifelse(reshaped$p.adjust.All < 0.01, paste0(temp, "**"), reshaped$beta.All)
reshaped$beta.All = ifelse(reshaped$p.adjust.All < 0.001, paste0(temp, "***"), reshaped$beta.All)

reshaped$`beta.Sex Interaction (Females - Males)` <- round(reshaped$`beta.Sex Interaction (Females - Males)`, 3)
temp <- reshaped$`beta.Sex Interaction (Females - Males)`
reshaped$`beta.Sex Interaction (Females - Males)` = ifelse(reshaped$`p.adjust.Sex Interaction (Females - Males)` < 0.05, paste0(temp, "*"), reshaped$`beta.Sex Interaction (Females - Males)`)
reshaped$`beta.Sex Interaction (Females - Males)` = ifelse(reshaped$`p.adjust.Sex Interaction (Females - Males)` < 0.01, paste0(temp, "**"), reshaped$`beta.Sex Interaction (Females - Males)`)
reshaped$`beta.Sex Interaction (Females - Males)` = ifelse(reshaped$`p.adjust.Sex Interaction (Females - Males)` < 0.001, paste0(temp, "***"), reshaped$`beta.Sex Interaction (Females - Males)`)

reshaped$beta.Female <- round(reshaped$beta.Female, 3)
temp <- reshaped$beta.Female
reshaped$beta.Female = ifelse(reshaped$p.adjust.Female < 0.05, paste0(temp, "*"), reshaped$beta.Female)
reshaped$beta.Female = ifelse(reshaped$p.adjust.Female < 0.01, paste0(temp, "**"), reshaped$beta.Female)
reshaped$beta.Female = ifelse(reshaped$p.adjust.Female < 0.001, paste0(temp, "***"), reshaped$beta.Female)

reshaped$beta.Male <- round(reshaped$beta.Male, 3)
temp <- reshaped$beta.Male
reshaped$beta.Male = ifelse(reshaped$p.adjust.Male < 0.05, paste0(temp, "*"), reshaped$beta.Male)
reshaped$beta.Male = ifelse(reshaped$p.adjust.Male < 0.01, paste0(temp, "**"), reshaped$beta.Male)
reshaped$beta.Male = ifelse(reshaped$p.adjust.Male < 0.001, paste0(temp, "***"), reshaped$beta.Male)

cat_order <- c("Regional Subcortical Volume", "Regional Cortical Volume","Global and Lobal Cortical Volume", "Regional Cortical Area", "Global and Lobal Cortical Area", "Regional Cortical Thickness","Global and Lobal Cortical Thickness", "Fractional Anisotropy", "Mean Diffusivity","General Fractional Anisotropy", "General Mean Diffusivity")

reshaped <- reshaped %>% 
  arrange(brain_phenotype_plot) %>%
  arrange(factor(brain_structure_type, levels = cat_order))

res_MDD <- reshaped %>%
  rename(`Brain Strucutre Phenotype` = brain_phenotype_plot,
         `Full Sample Beta` = beta.All,
         `Sex Interaction (Females - Males)` = `beta.Sex Interaction (Females - Males)`,
         `Female Sample Beta` = beta.Female,
         `Male Sample Beta` = beta.Male) %>%
  select(`Brain Strucutre Phenotype`, `Full Sample Beta`, `Sex Interaction (Females - Males)`, `Female Sample Beta`, `Male Sample Beta`)


reshaped <- reshape(res_disorder_interaction_brain_structure, idvar = c("brain_phenotype_plot", "brain_structure_type"), timevar = "sample", direction = "wide")
reshaped <- reshaped[grepl("brain_phenotype_plot|beta|p.adjust|brain_structure_type",names(reshaped))]

## Add astriks to significant betas
reshaped$beta.All <- round(reshaped$beta.All, 3)
temp <- reshaped$beta.All
reshaped$beta.All = ifelse(reshaped$p.adjust.All < 0.05, paste0(temp, "*"), reshaped$beta.All)
reshaped$beta.All = ifelse(reshaped$p.adjust.All < 0.01, paste0(temp, "**"), reshaped$beta.All)
reshaped$beta.All = ifelse(reshaped$p.adjust.All < 0.001, paste0(temp, "***"), reshaped$beta.All)

reshaped$`beta.Sex Interaction (Females - Males)` <- round(reshaped$`beta.Sex Interaction (Females - Males)`, 3)
temp <- reshaped$`beta.Sex Interaction (Females - Males)`
reshaped$`beta.Sex Interaction (Females - Males)` = ifelse(reshaped$`p.adjust.Sex Interaction (Females - Males)` < 0.05, paste0(temp, "*"), reshaped$`beta.Sex Interaction (Females - Males)`)
reshaped$`beta.Sex Interaction (Females - Males)` = ifelse(reshaped$`p.adjust.Sex Interaction (Females - Males)` < 0.01, paste0(temp, "**"), reshaped$`beta.Sex Interaction (Females - Males)`)
reshaped$`beta.Sex Interaction (Females - Males)` = ifelse(reshaped$`p.adjust.Sex Interaction (Females - Males)` < 0.001, paste0(temp, "***"), reshaped$`beta.Sex Interaction (Females - Males)`)

reshaped$beta.Female <- round(reshaped$beta.Female, 3)
temp <- reshaped$beta.Female
reshaped$beta.Female = ifelse(reshaped$p.adjust.Female < 0.05, paste0(temp, "*"), reshaped$beta.Female)
reshaped$beta.Female = ifelse(reshaped$p.adjust.Female < 0.01, paste0(temp, "**"), reshaped$beta.Female)
reshaped$beta.Female = ifelse(reshaped$p.adjust.Female < 0.001, paste0(temp, "***"), reshaped$beta.Female)

reshaped$beta.Male <- round(reshaped$beta.Male, 3)
temp <- reshaped$beta.Male
reshaped$beta.Male = ifelse(reshaped$p.adjust.Male < 0.05, paste0(temp, "*"), reshaped$beta.Male)
reshaped$beta.Male = ifelse(reshaped$p.adjust.Male < 0.01, paste0(temp, "**"), reshaped$beta.Male)
reshaped$beta.Male = ifelse(reshaped$p.adjust.Male < 0.001, paste0(temp, "***"), reshaped$beta.Male)

cat_order <- c("Regional Subcortical Volume", "Regional Cortical Volume","Global and Lobal Cortical Volume", "Regional Cortical Area", "Global and Lobal Cortical Area", "Regional Cortical Thickness","Global and Lobal Cortical Thickness", "Fractional Anisotropy", "Mean Diffusivity","General Fractional Anisotropy", "General Mean Diffusivity")

reshaped <- reshaped %>% 
  arrange(brain_phenotype_plot) %>%
  arrange(factor(brain_structure_type, levels = cat_order))

res_comobidity <- reshaped %>%
  rename(`Brain Strucutre Phenotype` = brain_phenotype_plot,
         `Full Sample Beta` = beta.All,
         `Sex Interaction (Females - Males)` = `beta.Sex Interaction (Females - Males)`,
         `Female Sample Beta` = beta.Female,
         `Male Sample Beta` = beta.Male) %>%
  group_by(`Brain Strucutre Phenotype`) %>%
  select(`Brain Strucutre Phenotype`, `Full Sample Beta`, `Sex Interaction (Females - Males)`, `Female Sample Beta`, `Male Sample Beta`)

```
```{r depression results, echo=F, warning=F}
flextable(res_MDD) %>%
  autofit() %>%
  fit_to_width(7) %>%
  bold(i = ~ grepl("\\*", res_MDD$`Full Sample Beta`), j = "Full Sample Beta") %>%
  bold(i = ~ grepl("\\*", res_MDD$`Sex Interaction (Females - Males)`), j = "Sex Interaction (Females - Males)") %>%
  bold(~ grepl("\\*", res_MDD$`Female Sample Beta`), j = "Female Sample Beta") %>%
  bold(~ grepl("\\*", res_MDD$`Male Sample Beta`), j = "Male Sample Beta")

```

Table S7 Full results for brain structure association analysis in current depression. \*, \*\* and \*\*\* represent significant results at p~FDR~\<.05, p~FDR~\<.01 and p~FDR~\<.001, respectively (FDR correction not applied to p-values from sex-interaction analysis). Significant results are highlighted in bold.

\newpage

```{r chronic pain results, echo=F, warning=F}

flextable(res_CP) %>%
  autofit() %>%
  fit_to_width(7) %>%
  bold(i = ~ grepl("\\*", res_CP$`Full Sample Beta`), j = "Full Sample Beta") %>%
  bold(i = ~ grepl("\\*", res_CP$`Sex Interaction (Females - Males)`), j = "Sex Interaction (Females - Males)") %>%
  bold(~ grepl("\\*", res_CP$`Female Sample Beta`), j = "Female Sample Beta") %>%
  bold(~ grepl("\\*", res_CP$`Male Sample Beta`), j = "Male Sample Beta")

```

Table S8 Full results for brain structure association analysis in current chronic pain. \*, \*\* and \*\*\* represent significant results at p~FDR~\<.05, p~FDR~\<.01 and p~FDR~\<.001, respectively (FDR correction not applied to p-values from sex-interaction analysis). Significant results are highlighted in bold.

\newpage

```{r CPMDD results, echo=F, warning=F}
flextable(res_comobidity) %>%
  autofit() %>%
  fit_to_width(7)  %>%
  bold(i = ~ grepl("\\*", res_comobidity$`Full Sample Beta`), j = "Full Sample Beta") %>%
  bold(i = ~ grepl("\\*", res_comobidity$`Sex Interaction (Females - Males)`), j = "Sex Interaction (Females - Males)") %>%
  bold(~ grepl("\\*", res_comobidity$`Female Sample Beta`), j = "Female Sample Beta") %>%
  bold(~ grepl("\\*", res_comobidity$`Male Sample Beta`), j = "Male Sample Beta")

```

Table S9 Full results for chronic pain interaction and sex-chronic pain interaction on the association of depression and brain structure. \*, \*\* and \*\*\* represent significant results at p~FDR~\<.05, p~FDR~\<.01 and p~FDR~\<.001, respectively, respectively (FDR correction not applied to p-values from sex-interaction analysis). Significant results are highlighted in bold.
\newpage
